+++ references.bib
@@ -348,6 +348,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
                apply the introduced method in modelling of anisotropic
                potentials and to be used when twisting between layers is
                present.",
+  DOI	  = {10.1088/2399-6528/aba035},
   journal   = "J. Phys. Commun.",
   publisher = "IOP Publishing",
   volume    =  4,
