### 2024-11-19 #7458 @arkavo
> * [ ]  Please study the above reference check ☝️ and see if you can address any of the reported potential DOI issues. You can add/amend DOI entries in your `.bib` file, and call `@editorialbot check references` here to check them again.
> * [ ]  Your project appears to lack contributing guidelines at the moment. Please work to add these and to mention/link to them in your README. You could for instance create a `CONTRIBUTING.md` file (see here for some examples: https://contributing.md/example/).

@Kevin-Mattheus-Moerman I've added the necessary missing reference and the `contributing.md` and `code_of_conduct.md` files. Please move forward as necessary
Thanks.
