### 2024-11-20 #7458 @phibeck
Hi @arkavo, thanks for your submission. I'll be looking for reviewers next. When you have a moment, it would be helpful if you could:
- identify a few potential reviewers from this list https://reviewers.joss.theoj.org/reviewers (without tagging them with an @)
- please check the rendering of the manuscript, as there seem to be issues with equations, numbers and symbols
- think about whether you can shorten the summary in the paper to avoid duplication of the software documentation in the git repo. I see that the figures and part of the summary is already contained in the docs, so perhaps it makes sense to move the rest of the summary there as well, and instead include a "summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience" (see https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain) As a side note, your paper is on the longer side with >2000 words (we aim for 250-1000 words)
