### 2024-11-21 #7458 @arkavo
Hi @phibeck

- We have identified 2 potential reviewers from the JOSS list as Xu He and Hugo Strand which we think would be an ideal fit for our work.
- We are working on the other points that you have raised regarding the manuscript.
