### 2024-12-11 #7458 @ChrisWoodgate
@phibeck Apologies for a slightly delayed response—I wanted to check that I still had access to a workstation with GPU capabilities where I could run interactively. (Fortunately, it turns out that I do!)

I should be able to review this, yes. A little busy this week but I would anticipate being able to return comments this side of Christmas
