### 2024-12-13 #7589 @editorialbot
[REVIEW]: CUDA-METRO: Parallel Metropolis Monte-Carlo for 2D Atomistic Spin Texture Simulation
**Submitting author:** @arkavo (<a href="http://orcid.org/0009-0006-6741-9377">Arkavo Hait</a>)
**Repository:** https://github.com/arkavo/CUDA-METRO
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.8
