### 2024-12-13 #7589 @tpurcell90
I did the initial repo checks and did not find the scripts used to recreate the figures did I miss them somewhere @arkavo ?

Also when doing the review where would you want me to place any documentation issues here, in the issue tracker of the repo, or both?
