### 2024-12-14 #7589 @arkavo
Hi, @tpurcell90. Thanks for taking the time to review. We don't have the scripts for any figures since we didn't know we would specifically need them. Since they are required, we will regenerate these figures and attach a script for them very soon.
