### 2024-12-16 #7589 @arkavo
Also, I think we should place the documentation issues here itself. It should be fine.

Additionally, I saw that I had given the scripts for fig 1 and fig 2 but they're not obvious at first glance. Should I modify the main page with context on how to arrive at the diagrams or is a separate script necessary?

Thanks
