+++ placeholder.md
@@ -190,3 +190,6 @@ Board (SERB), Government of India, under Grant No. CRG/2020/000758.
         \EndProcedure
     \end{algorithmic}
 \end{algorithm}
+
+ - \usepackage[utf8]{inputenc}
+ - \usepackage{lmodern}
