### 2024-12-17 #7589 @arkavo
@tpurcell90 I added 3 scripts under ```/src```. Please see if they work as intended on your end. They were working for me.
