### 2025-01-07 #7589 @phibeck
Hi @tpurcell90 thanks for getting your review started! :tada: Please let us know whether the scripts provided by @arkavo work as intended, and how the rest of your review is going.

@ChrisWoodgate thanks again for reviewing! Please go ahead and create your checklist first using the command `@editorialbot generate my checklist`. As indicated by the authors, please document any issues as comments here. Please let me know if you have any questions about the review process.
