### 2025-01-08 #7589 @tpurcell90
Apologies for the delay, but apparently my nvidia drivers broke sometime over the last two months so I had to set this up on the HPC to test it.

When trying to run the `main.py` script or `fig1_script.ipynb` it runs into path not found errors when not run in the `src` directory because the package does not actually install from what I can tell (no build instruction in `pyproject.toml` file). Even running it in the `src` directory fails because of another hard coded path to the materials in `src/construct.py`

```
77 self.MAT_NAME, self.MAT_PARAMS = rm.read_2dmat("../../"+self.Input_Folder+"TC_"+self.Material+".csv")
```
When running in src this should be `"../"+self.Input_Folder+"TC_"+self.Material+".csv"`

Would it be possible to install this as a module and remove the hard coded paths? Once that is done I can try the rest of the package.
