### 2025-01-09 #7589 @arkavo
Hi @tpurcell90. Thanks for reviewing.

I'm looking into the paths issue right now to fix it.
What do you mean by "install as a module"? something like you can "pip install" or something else?

Additionally, the paper issues are only occurring with the JOSS compiler for some reason, in any local compiler it works fine, even in the VSCode extension, any suggestion or help on this matter would be greatly appreciated. In the meantime, the readme.md should have the main equation rendered.
