### 2025-01-09 #7589 @tpurcell90
> I'm looking into the paths issue right now to fix it.
> What do you mean by "install as a module"? something like you can "pip install" or something else?

Ya allow your code to be installed using `pip install .` where that installs the module into `site-packages` so you can run the code from any directory without modifying the `PYTHONPATH` environment variable.

For the path issue I would recommend adding it as an input parameter with a default path set relative to the absolute path to an installed module.

>Additionally, the paper issues are only occurring with the JOSS compiler for some reason, in any local compiler it works fine, even in the VSCode extension, any suggestion or help on this matter would be greatly appreciated. In the meantime, the readme.md should have the main equation rendered.

I would look into the resources here, I had similar issues that I fixed by locally compiling the paper using their services, and figuring out what was wrong. https://joss.readthedocs.io/en/latest/paper.html#checking-that-your-paper-compiles
