+++ pyproject.toml
@@ -3,8 +3,8 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "cuda-metro-line"
-version = "0.0.1"
+name = "cuda-metro"
+version = "1.0.0"
 authors = [
   { name="Arkavo Hait", email="arkavohait@iisc.ac.in" },
   { name="Santanu Mahapatra", email="santanu@iisc.ac.in"}
