### 2025-01-19 #7589 @ChrisWoodgate
Apologies for the delay on returning comments. I am only just back from a conference. Regarding the paper, I have a few comments, which echo those @tpurcell90 has already raised.

Paper comments:

- [ ] In the draft PDF version of the manuscript, many of the equations are currently not being rendered (or being rendered incorrectly). See as an example the attached image of how the first equation of the summary section is being rendered for me. (There are currently lots of these rendering issues for me, which makes it difficult to complete this part of the review at present.)
- [ ] Similarly, the code blocks in the main text don't appear to be rendering correctly for me. (See the second attached screenshot.)

It would be great to have these resolved to make finishing the 'paper' section of the review less challenging.
