### 2025-01-19 #7589 @ChrisWoodgate
### Documentation Comments

- [ ] I think it would be really helpful if, in the README, there was a step-by-step guide for how to run an example calculation and visualise the results. So far as I can tell, what I need to do is navigate to the source directory and execute `python3 main.py`, which I can do successfully. But how do I then view the results of the simulation? This could be made much clearer, I feel.
