### 2025-01-19 #7589 @ChrisWoodgate
### Functionality Comments

- [ ] If I am to run `main.py` in the `src` directory, I need to modify line 10. The reference `../configs/test_config.json` should actually be `../../configs/test_config.json`, unless I am mistaken?
