### 2025-01-20 #7589 @phibeck
Thanks @tpurcell90 and @ChrisWoodgate for your updates. The rendering issue was already present at the [pre-review] which I had also pointed out to the authors then, but I forgot to double-check before starting the [review] issue. @arkavo could you please take care of this so that the reviewers can continue with their assessment of the manuscript? As @tpurcell90 pointed out, the easiest way to figure out the issue is to follow the instructions given in the link, compiling it locally using the same routines.

@tpurcell90:
> Paper comments
> - [@phibeck](https://github.com/phibeck)  Editorial question: The paper is showing new scientific results in it, but more as an example of the capabilities of the code. Is this okay within the guidelines of JOSS?

Thank you for this. The guidelines of JOSS state: `Your paper must not focus on new research results accomplished with the software.` (https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) Indeed, since our reviewers focus on reviewing the software and not the scientific new results, JOSS papers should be cited for the implementation not for the results. @arkavo could you please comment on the novelty of the results presented in the manuscript and whether you intend to publish them in field-specific journal?
