+++ .github/workflows/publish.yml
@@ -38,7 +38,7 @@ jobs:
   pypi-publish:
     runs-on: ubuntu-latest
     needs:
-      - release-build
+      - build
     permissions:
       # IMPORTANT: this permission is mandatory for trusted publishing
       contents: read
