+++ .github/workflows/publish.yml
@@ -37,8 +37,7 @@ jobs:
 
   pypi-publish:
     runs-on: ubuntu-latest
-    needs:
-      - build
+    
     permissions:
       # IMPORTANT: this permission is mandatory for trusted publishing
       contents: read
