+++ .github/workflows/publish.yml
@@ -37,7 +37,8 @@ jobs:
 
   pypi-publish:
     runs-on: ubuntu-latest
-    
+    needs:
+      - release-build
     permissions:
       # IMPORTANT: this permission is mandatory for trusted publishing
       contents: read
@@ -65,3 +66,4 @@ jobs:
         uses: pypa/gh-action-pypi-publish@release/v1
         with:
           packages-dir: dist/
+          password: ${{ secrets.TEST_PYPI_API_TOKEN }}
