+++ src/cudametro/__init__.py
@@ -1,1 +1,4 @@
-from construct import construct as cm
+import sys
+#import montecarlo as montecarlo
+from .construct import MonteCarlo
+from .construct import Analyze
