+++ src/cudametro.egg-info/SOURCES.txt
@@ -1,0 +1,23 @@
+LICENSE
+README.md
+pyproject.toml
+src/cudametro/Material_Reader.py
+src/cudametro/__init__.py
+src/cudametro/bench.py
+src/cudametro/construct.py
+src/cudametro/curie_temp.py
+src/cudametro/engraph.py
+src/cudametro/grid_bench.py
+src/cudametro/main.py
+src/cudametro/montecarlo.py
+src/cudametro/visualize.py
+src/cudametro.egg-info/PKG-INFO
+src/cudametro.egg-info/SOURCES.txt
+src/cudametro.egg-info/dependency_links.txt
+src/cudametro.egg-info/top_level.txt
+src/cudametro/deprecated/alt_mat_1_construct.py
+src/cudametro/deprecated/cri3_analysis.py
+src/cudametro/deprecated/dmi.py
+src/cudametro/deprecated/double_mat.py
+src/cudametro/deprecated/sub_construct.py
+src/cudametro/deprecated/test_cpu.py
