+++ pyproject.toml
@@ -1,10 +1,11 @@
 [build-system]
-requires = ["setuptools>=75.0"]
-build-backend = "setuptools.build_meta"
-
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+[tool.setuptools.packages.find]
+namespaces=false
 [project]
 name = "cudametro"
-version = "1.0.0"
+version = "1.1.1"
 authors = [
   { name="Arkavo Hait", email="arkavohait@iisc.ac.in" },
   { name="Santanu Mahapatra", email="santanu@iisc.ac.in"}
