### 2025-01-26 #7589 @arkavo
Packaging the code is done with the same functionality, I can import and add an entire new script. It's on PyPI as ```cudametro-1.1.1```. If it is still inaccessible on your end, please let me know.

Thanks.
