### 2025-02-01 #7589 @phibeck
> Hi [@phibeck](https://github.com/phibeck), [@tpurcell90](https://github.com/tpurcell90) and [@ChrisWoodgate](https://github.com/ChrisWoodgate), I am trying to get the packaging sorted out as soon as possible and looking into the paper compilation simultaneously. Fortunately, the paper compiles locally but without the JOSS flairs so I will put a vanilla version of the paper in the main directory as a .pdf as a temporary issue till I figure out the other issue.

Thank you for this, that should be okay as a workaround. But please try to make sure the paper compiles correctly as soon as possible, since that makes it easier to track in case of any changes. @tpurcell90 and @ChrisWoodgate please let us know if you have other issues with the formatting or suggestions for the manuscript.

> Regarding the issue of "new results", I would say the results (figures) given in the paper are all previous existing results which are also source-linked to the figures themselves. We are demonstrating the applicability of these existing results in our code. The only "new" result we showed was a simulation from lifetime to death of a structure which demonstrated the code's speed. If this is not clear do let me know and I will change the language result presentation. Hope this helps and thanks again for helping out.

Thanks for the clarification. That sounds reasonable to me as long as the focus is on the speedup rather than the new scientific content. @tpurcell90 do you have suggestions for possible modifications of the text to clarify this if needed? Thanks
