### 2025-02-13 #7589 @arkavo
Hey @phibeck. So I checked the paper compilation from my end and can see 2 things:

1. The automatic workflow from the github actions compile the artefact paper with the symbols all intact

![Image](https://github.com/user-attachments/assets/159f6033-4503-4773-a230-eb5591b667d7)

even in the algorithm part as shown (this is from the official JOSS workflow)

![Image](https://github.com/user-attachments/assets/5e62172a-a011-4601-a5a2-dea560f9648c)

2. When I compiled the paper locally using pandoc@latest, it *still* compiles with all symbols intact (but of course without the watermark or decorators)

Is there a possibility that the problem persists on your end? The precompiled pdf also exists in the repo home till this is sorted fully.

The code is now fully packaged under pypi as cudametro-1.1.1 and anyone can install it via pip with the other libraries and import. The documentation remains the same(regarding the API) but I can add more stuff if needed.

Functionality should be fine (it still works for me) but if any of the reviewers have any difficulty running and or installing, I would be more than happy to assist.

Thanks.
