### 2025-02-17 #7589 @tpurcell90
When I try to run the fig 1 notebook I get the following error
```
LogicError                                Traceback (most recent call last)
Cell In[1], line 15
     13 test_mc0.grid_reset()
     14 test_mc0.generate_random_numbers(test_mc0.S_Wrap)
---> 15 m, x = test_mc0.run_mc_tc_3636(t)
     16 M = np.append(M, m)
     17 X = np.append(X, x)

File ~/.conda/envs/cuda_metro/lib/python3.11/site-packages/cudametro/construct.py:288, in MonteCarlo.run_mc_tc_3636(self, T)
    286 drv.memcpy_htod(self.BJ, beta)
    287 for j in range(self.stability_runs):
--> 288     mc.METROPOLIS_MC_DM0_3_6_3_6(self.GPU_MAT, self.GRID_GPU, self.BJ, self.NFULL[j*self.Blocks:(j+1)*self.Blocks-1], self.S1FULL[j*self.Blocks:(j+1)*self.Blocks-1], self.S2FULL[j*self.Blocks:(j+1)*self.Blocks-1], self.S3FULL[j*self.Blocks:(j+1)*self.Blocks-1], self.RLIST[j*self.Blocks:(j+1)*self.Blocks-1], self.GPU_TRANS, self.B_GPU, self.GSIZE, block=(self.Threads,1,1), grid=(self.Blocks,1,1))
    289     mc.GRID_COPY(self.GRID_GPU, self.GPU_TRANS, block=(1,1,1), grid=(self.Blocks,1,1))
    290 drv.memcpy_dtoh(self.grid, self.GRID_GPU)

File ~/.conda/envs/cuda_metro/lib/python3.11/site-packages/pycuda/driver.py:481, in _add_functionality.<locals>.function_call(func, *args, **kwargs)
    478 if block is None:
    479     raise ValueError("must specify block size")
--> 481 func._set_block_shape(*block)
    482 handlers, arg_buf = _build_arg_buf(args)
    484 for handler in handlers:

LogicError: cuFuncSetBlockShape failed: an illegal memory access was encountered
```

What system configuration have you tried to run this on?

Also it would be good to remove all hard-coded paths with `../../` in it, and just use the `__file__` keyword to specify a base directory if needed.
