### 2025-02-18 #7589 @phibeck
Hi @arkavo it looks like your changes fixed the compilation issues, the article proof looks okay to me now. Can you confirm that everything is rendered correctly?
