### 2025-02-18 #7589 @ChrisWoodgate
Okay, I have taken a look at the revised paper and looked back at the comments/issues I have raised previously. There is one new comment I would make regarding the paper, which relates to the typesetting of elemental names.

- [ ] At the moment, _e.g._ $CrI_3$ is typeset all in italics like a mathematical variable. I think it would be better to use the `\textrm` command to render these elemental names not in italics. _E.g._ $\textrm{CrI}_3$, as this is convention in most materials physics journals.

The remaining issues (copy-pasted from earlier comments) I have for @arkavo are:

## Paper comments

- [ ] Regarding the state of the field. At the moment I can't see mention of other packages for performing magnetic texture simulation and how CUDA-METRO differs from these. The two main codes I can think of in this domain are VAMPIRE (University of York) and Spirit (Forschungszentrum Jülich). It would be good to compare with these existing codes. Relevant literature is cited but I do not see any comparison of functionalities.

## Documentation comments

- [ ] I think it would be really helpful if, in the README, there was a step-by-step guide for how to run an example calculation and visualise the results. So far as I can tell, what I need to do is navigate to the source directory and execute python3 main.py, which I can do successfully. But how do I then view the results of the simulation? This could be made much clearer, I feel.

## Functionality comments

- [ ] If I am to run main.py in the src directory, I need to modify line 10. The reference ../configs/test_config.json should actually be ../../configs/test_config.json, unless I am mistaken?

Perhaps some of these issues have already been addressed? But I don't see replies to my earlier comments so it is not clear to me whether or not I can 'close' these issues. Thanks!
