### 2025-02-18 #7589 @tpurcell90
The paper looks good to me, the only possible thing missing is the State of the Field section.

Also I can run main.py, but not the fig1 notebook if that helps give you an idea of what is happening with that config.
