### 2025-02-18 #7589 @tpurcell90
I concur having some sort of start to end tutorial would be good.

Also it looks like the default output directory is set to `../../outputs` I personally think this can be dangerous as it is unknown which directory it will be run in. In theory `../../` may not even exist, even though this would be bad practice.
