### 2025-02-19 #7589 @arkavo
> Hi [@arkavo](https://github.com/arkavo) it looks like your changes fixed the compilation issues, the article proof looks okay to me now. Can you confirm that everything is rendered correctly?

@phibeck looks good. I'm working on @tpurcell90 and @ChrisWoodgate's comments in the meantime.
