+++ paper.md
@@ -109,13 +109,13 @@ Unlike the SSU scheme, we do not extract simulation data after every $N^2$ spin 
 
 We validate our algorithm by accurately reproducing temperature-dependent magnetic phase transitions and intricate topological spin textures already reported in the literature, either by SSU based Monte Carlo or by Landau Lifshitz Gilbert(LLG)[@gilbert_classics_2004] spin dynamics, for diverse crystal structures.
 
-It was once thought that long-range magnetic order could not exist [@PhysRevLett.17.1133] in two or one-dimensional materials. In 2017, $CrI_3$ was one of the first 2D materials [@Huang2017] where ferromagnetism was demonstrated. Using CUDA-METRO, we first simulate the magnetic phase transition of $CrI_3$ from ferromagnetic to paramagnetic with increasing temperature. In Fig 1, we reproduce magnetic phase transition from [@Kartsev2020], and show the effect of parallelization with respect to the reference data. As mentioned before, we can obtain very accurate results with respect to SSU methods by limiting the parallelization at $10\%$ or below of $N^2$. The ferromagnetic to paramagnetic transition point is known as Critical temperature and is extracted from the peak of the susceptibility versus temperature plot.
+It was once thought that long-range magnetic order could not exist [@PhysRevLett.17.1133] in two or one-dimensional materials. In 2017, \textrm{CrI_3} was one of the first 2D materials [@Huang2017] where ferromagnetism was demonstrated. Using CUDA-METRO, we first simulate the magnetic phase transition of \textrm{CrI_3} from ferromagnetic to paramagnetic with increasing temperature. In Fig 1, we reproduce magnetic phase transition from [@Kartsev2020], and show the effect of parallelization with respect to the reference data. As mentioned before, we can obtain very accurate results with respect to SSU methods by limiting the parallelization at $10\%$ or below of $N^2$. The ferromagnetic to paramagnetic transition point is known as Critical temperature and is extracted from the peak of the susceptibility versus temperature plot.
 
 Next, we demonstrate the nucleation of topological spin textures which are emergent phenomena in condensed matter physics and are rapidly having importance in information technology[@luo_skyrmion_2021;@bessarab_lifetime_2018]. While MC simulations of medium-sized supercells $(64 \times 64)$ yield quite accurate results for Critical temperature calculation, observing topological spin textures, once needs much larger supercells.
 
 # Results
 
-First, we simulate skyrmions in $MnBr_2$[@acs_nanolett] as shown in Fig 2. $MnBr_2$ is a square lattice and for this simulation, we have considered up to the second nearest neighbour. This material exhibits anisotropic DMI with an anti-ferromagnetic ground state. An anti-ferromagnetic skyrmion spin texture is accurately reproduced in our simulation. Anti-ferromagnetic skyrmions are technologically important since they do not exhibit skyrmion Hall effect. We further study the material $CrInSe_3$ [@du_spontaneous_2022] which has a hexagonal lattice. This simulation was conducted considering only the nearest neighbours and the formation of skyrmions is shown in Fig 2. Once again our results are in agreement with the original report. All these simulations were conducted in a $200 \times 200(49 \times 49nm^2)$ supercell and took 30s to stabilize these topological spin textures at a parallelization of $20\%$ conducted on a V100-SXM2 processor.
[…]
