### 2025-02-20 c13fc18
textrm to chemformula
