+++ paper.md
@@ -28,7 +28,7 @@ header-includes:
  - \usepackage{algorithm}
  - \usepackage[noend]{algpseudocode}
  - \usepackage{amsmath,amssymb}
- - \usepackage{chemformula}
+ - \usepackage[version=4]{mchem}
 date: 30 September 2024
 bibliography: references.bib
 
