### 2025-02-20 ac797ca
brute forcing formulas
