+++ paper.md
@@ -130,7 +130,7 @@ Fig 3: Lifetime of a skyrmion in $\text{MnSTe}$, from its creation to annihilati
 
 # State of Field
 
-THe state of micromagnetics have been moderately studied since the advent of computers with the power to compute. To this effect, some of the notable existing software would be [VAMPIRE](https://vampire.york.ac.uk/) and [SPIRIT](https://spirit-code.github.io/). While SPIRIT uses an LLG based approach to calculate the electron spin vectors, our code uses a Monte Carlo based approach, which aids in extended sampling of the phase space. VAMPIRE uses a Monte Carlo based approach but is primarily a CPU(C++ MPI) based code which is outdated by today's standards due to the advent and mass usage of the GPU architecture. CUDA-METRO was designed and written to achieve a modular GPU-based framework which would work well on modern hardware and use a more efficient approach the exploring the phase space of micromagnetics. 
+THe state of micromagnetics have been moderately studied since the advent of computers with the power to compute. To this effect, some of the notable existing software would be [VAMPIRE](https://vampire.york.ac.uk/)[@evans_atomistic_2014] and [SPIRIT](https://spirit-code.github.io/)[@muller_spirit_2019]. While SPIRIT uses a single spin approach to calculate the electron spin vectors, our code uses a multi spin based approach, which aids in extended sampling of the phase space. VAMPIRE uses a Monte Carlo based approach but is primarily a CPU(C++ MPI) based code which is outdated by today's standards due to the advent and mass usage of the GPU architecture. CUDA-METRO was designed and written to achieve a modular GPU-based framework which would work well on modern hardware and use a more efficient approach the exploring the phase space of micromagnetics.
 
 # Acknowledgements
 This work is supported by the Core Research Grant (CRG) scheme of the Science and Engineering Research
