### 2025-02-21 65a54ae
Updated text and binds
