+++ paper.md
@@ -27,6 +27,7 @@ affiliations:
 header-includes:
  - \usepackage{algorithm}
  - \usepackage[noend]{algpseudocode}
+ - \usepackage{float}
  - \usepackage{amsmath,amssymb}
 date: 30 September 2024
 bibliography: references.bib
@@ -59,7 +60,7 @@ Where $J$ is the isotropic exchange parameter, the $K$s are the anisotropic exch
 
 Starting from a random spin configuration, in this many-body problem, our objective is to find the orientation of spin vectors for every atom so that the energy of the entire lattice reaches to its minimum for a given magnetic field and temperature.
 Traditionally single spin update(SSU) scheme is employed to solve this problem, which satisfies the detailed balance condition. In the SSU method of updating the state, a single atomic spin is chosen at random and changed, while noting down the energy shift. This new state is then accepted or rejected using the Metropolis criteria as shown in Algorithm 1, where $\beta=(k_bT)^{-1}$, $k_b$ being the Boltzmann constant and $T$ being the temperature. It is imperative that SSU becomes extremely inefficient as the dimensionality increases.
-\begin{algorithm}[t]
+\begin{algorithm}[H]
     \caption{Metropolis Selection}
     \label{algorithm:MS1}
     \begin{algorithmic}[0]
@@ -75,7 +76,7 @@ Traditionally single spin update(SSU) scheme is employed to solve this problem, 
     \end{algorithmic}
 \end{algorithm}
 In our method, as depicted in Algorithm 2, we select multiple atomic spins at the same time and change them all at once, treating them as independent events. For any individual spin, they do not feel the effects of the other changed spins. In each of these points, we use the Metropolis criteria to accept or reject the changed spin vectors. This becomes our new state. Here $P$ denotes the number of lattice points we are evaluating at the same time for any given state, while $\Gamma$ is the batch size. Tuning $\Gamma$ ensures that we can fill up our VRAM with pre-generated random numbers instead of generating $4\times P$ numbers per step. These 4 random number arrays are further processed into $n$, our site selection, $(\theta,\phi)$, which become the angle coordinates for a new random spin vector and $r$ which is a conditional uniform random number used to evaluate the Metropolis criteria.
-\begin{algorithm}[t]
+\begin{algorithm}[H]
     \caption{Parallel Monte Carlo}
     \label{algorithm:MS2}
     \begin{algorithmic}[0]
