+++ README.md
@@ -7,12 +7,10 @@ A pyCUDA based Metropolis Monte Carlo simulator for 2D-magnetism systems. This c
 To use, clone the repository using
 
 ```
-
 git clone https://github.com/arkavo/CUDA-METRO.git
-
 ```
 
-into a directory of choice. Creating a new python3 environment is recommended to run the code.
+into a directory of choice. Creating a new python3(>=3.8) environment is recommended to run the code.
 
 To setup the environment after creation, simply run 
 				```pip install -r requirements.pip```
@@ -21,9 +19,47 @@ Some template codes have already been provided under folder ```/src``` along wit
 
 There are 4 seperate files to design your simulations. To run a simple simulation, use ```python main.py <input_file_name>```.
 
+# Executing a simulation
+This is a standalone code with ```pyCuda``` backend. An "Input" to this code's kernel would be a matrix which has 3D spins of lattices as an array with the structure ```[site1_spin_x site1_spin_y site1_spin_z site2_spin_x ..... siteN_spin_z]```. Executing the simulation will then access this array and perform a Metropolis Monte Carlo simulation which will then give an output which will be automatically saved in the save directory with the same array structure in a ```.npy``` format.
+
+## Setup
+To execute a simulation, you will need 3 files.
+<list>
+<li>Config file (examples in /configs/test_config.json)
+<li>Material parameters (examples in /inputs)
+<li>Script file (examples in /src/cudametro)
+</list>
+
+The ```Config file``` determines the parameters for the simulation, including the amount of VRAM used by your GPU if you are facing crashes or out of memory issues.
+
+The ```Material parameters``` file contains all the properties of the material(interaction parameters) as a vector. The crystal configuration is also stored in this file.
+<details>
+<summary>Vector</summary>
+<br>
+[name, spin, J1, J2, J3, J4, K1x, K1y, K1z, K2x, K2y, K2z, K3x, K3y, K3z, K4x, K4y, K4z, Ax, Ay, Az, Tc(experimental), structure, DMI]
+</details>
+
+## Execution
+
+To execute a fresh simulation
+
+run the ```script file``` with an ```input file```. The script file must have the construct library bindings as defined in [Construct MonteCarlo](#construct-montecarlo). This will also generate an output folder which contains the lattice of spins in the format ```[site1_spin_x site1_spin_y site1_spin_z site2_spin_x ..... siteN_spin_z]``` with $N=n^2$, the last spin. The contents of this folder can be easily visualized by using the bindings of the provided [Construct Analyze](#construct-analyze) library. Note that a seperate script has to be wrtitten. A simple sample of such a script is given in ```/src/visualize.py``` which is ready to run and can analyze any given folder.
+
+To execute a simulation, run
+```python <script> <input>```
+
+To visualize a simulation, run
+```python <visual> <folder_path>```
+if you are running a custom job but simply running
+```python visualize.py <folder_path>``` will work for most cases.
+
+To run a Critical temperature analysis, execute ```python tc_sims.py``` after configuring the appropriate config file in ```/configs/tc_config.json```. The script will create its own graph at runtime, no additional script needed. If facing out of memory error, please consider lowering "Block" count in the config file.
+
 # Custom input files
 
-If one wants to create their own input file from scratch, just copy this json template at your convenience:
+##Config
+
+If one wants to create their own ```config``` file from scratch, please copy this json template:
 ```
 {
     "Single_Mat_Flag" : 1 if single material 0 otherwise (int),
[…]
