+++ README.md
@@ -139,21 +139,23 @@ The template file to run a visual analyzer is given in ```visualize.py```, this 
 
 We consider a lattice system with a periodic arrangement of atoms, where each atom is represented by a 3D spin vector.  This atomistic spin model is founded on the spin Hamiltonian, which delineates the essential spin-dependent interactions at the atomic scale, excluding the influences of potential and kinetic energy and electron correlations. The spin Hamiltonian is conventionally articulated as:
 
-$$
+```math
+
 H_i=  -\sum_j J_1s_i\cdot s_j - \sum_j K^x_1 s^x_i s^x_j-\sum_j K^y_1 s^y_i s^y_j-\sum_j K^z_1 s^z_i s^z_j-\sum_k J_2 s_i\cdot s_k - \sum_k K^x_2 s^x_i s^x_k
-$$
 
-$$
+\\
+
 -\sum_k K^y_2 s^y_i s^y_k -\sum_k K^z_2 s^z_i s^z_k -\sum_l J_3s_i\cdot s_l - \sum_l K^x_3 s^x_i s^x_l-\sum_l K^y_3 s^y_i s^y_l-\sum_l K^z_3 s^z_i s^z_l 
-$$
 
-$$
+\\
+
 -\sum_m J_4s_i\cdot s_m - \sum_m K^x_4 s^x_i s^x_m-\sum_m K^y_4 s^y_i s^y_m -\sum_m K^z_4 s^z_i s^z_m - A s_i \cdot s_i-\sum_j \lambda(s_i\cdot s_j)^2  
-$$
 
-$$
+\\
+
 -\sum_j D_{ij}\cdot (s_i \times s_j) -\mu B \cdot s_i
-$$
+
+```
 
 Where $J$ is the isotropic exchange parameter, the $K$s are the anisotropic exchange parameters, with the superscript denoting the spin direction, $A$ is the single ion exchange parameter, $\lambda$ is the biquadratic parameter, $D$ is the Dyzaloshinskii-Moriya Interaction(DMI) parameter. $\mu$ is the dipole moment of a single atom and $B$ is the external magnetic field. $s_i,s_j$ are individual atomic spin vectors. $\{s_j\}$ are the first set of neighbours, $\{s_k\}$ are the second set of neighbours and so on. The subscripts below all $J$s and $K$s denote the neighbour set, $J_1$ denotes the first neighbours, $J_2$ the second and so on. From this equation, we see that energy of an atom depends on its interactions with the neighbours. In our code, we have limited the number of neighbour sets to be 4 since it is expected for 2D materials that the interaction energy dies down beyond that. All these above parameters except $B$ are material specific parameters that are the inputs to our MC code.
 
@@ -173,7 +175,7 @@ In Fig 5, we further show a similar life cycle evolution for a giant skyrmion of
 
 |![Figure 2](figures/Figure_2.png)|
 |:--:|
-| *Fig 2: Presence of anti-merons and merons in CrCl<sub>3. The color bar represents normalized spin vectors in the z direction.* |
+| *Fig 2: Presence of anti-merons and merons in CrCl<sub>3</sub>. The color bar represents normalized spin vectors in the z direction.* |
 
 |![Figure 3](figures/Figure_3.png)|
 |:--:|
