### 2025-02-25 15bc45e
1.1.2 ver
better tutorial section
clearer docu
renamed "working principle" to "theory"
fig1 size reduced by 2x (all parallels by 4x)
fig2 size reduced by 2x (all parallels by 4x)
fig3 size reduced by 2x (all parallels by 4x)

renamed main.py -> sample.py
incorporated visualizer in sample.py
