### 2025-03-03 #7589 @phibeck
Hi @arkavo thanks for the update, that indeed seems to solve the issue with the equation being split.

@ChrisWoodgate and @tpurcell90 when you have a moment, could you take a look and see whether these updates address your previous issues/suggestions adequately? Thank you.
