### 2025-03-09 351871c
added alternate scripts
