+++ src/cudametro/curie_temp.py
@@ -2,7 +2,7 @@
 #--------------------------------------------
 
 import construct as cst
-import montecarlo as mc
+import cudametro.montecarlo as mc
 import numpy as np
 import matplotlib.pyplot as plt
 
