+++ src/cudametro/bench.py
@@ -2,7 +2,7 @@
 
 #--------------------------------------------
 import construct as cst
-import montecarlo as mc
+import cudametro.montecarlo as mc
 import numpy as np
 import tqdm as tqdm
 import datetime as dt
