+++ src/cudametro/grid_bench.py
@@ -1,5 +1,5 @@
 import construct as cst
-import montecarlo as mc
+import cudametro.montecarlo as mc
 import numpy as np
 import tqdm as tqdm
 import datetime as dt

[…]
