+++ sample_scripts/fig1_script.py
@@ -6,57 +6,76 @@ import tqdm as tqdm
 # =============================================================================
 # Edit the following lines to change the parameters of the simulation
 # =============================================================================
-fname = "fig1_configs/p2.json"          # Configuration file [options p1.json, p2.json, p3.json, p4.json]
+fname = "fig1_configs"          # Configuration file [options p1.json, p2.json, p3.json, p4.json]
 in_dir = "input_parameters/"                             # Input directory
 # =============================================================================
 
 
-test_mc0 = cst.MonteCarlo(config=fname, input_folder=in_dir)
-test_mc0.mc_init()
-test_mc0.display_material()
-M, X = np.array([]), np.array([])
-for t in test_mc0.T:
-    test_mc0.grid_reset()
-    test_mc0.generate_random_numbers(test_mc0.S_Wrap)
-    m, x = test_mc0.run_mc_tc_3636(t)
-    M = np.append(M, m)
-    X = np.append(X, x)
-np.save(f"{test_mc0.save_directory}/BQM_{test_mc0.size}_{test_mc0.Blocks}", M)
-np.save(f"{test_mc0.save_directory}/BQX_{test_mc0.size}_{test_mc0.Blocks}", X)
+# Universal parameters
+save_dir_flag = False
+T_arr = None
+spin = 1.0
+mat = None
+size = 32
 
-X[0] = 0.0
+for i in range(4):
+    fs_name = f"fig1_configs/p{i+1}.json"
+    print(f"Running {fs_name}")
+    test_mc0 = cst.MonteCarlo(config=fs_name, input_folder=in_dir)
+    test_mc0.mc_init()
+    test_mc0.display_material()
+    M, X = np.array([]), np.array([])
+    for t in test_mc0.T:
+        test_mc0.grid_reset()
+        test_mc0.generate_random_numbers(test_mc0.S_Wrap)
+        m, x = test_mc0.run_mc_tc_3636(t)
+        M = np.append(M, m)
+        X = np.append(X, x)
+    if not save_dir_flag:
+        unified_save_dir = test_mc0.save_directory
+        save_dir_flag = True
+        save_dir = test_mc0.save_directory
+        T_arr = test_mc0.T
+        spin = test_mc0.spin
+        mat = test_mc0.MAT_NAME
+        size = test_mc0.size
+    np.save(f"{unified_save_dir}/BQM_{test_mc0.size}_{test_mc0.Blocks}", M)
+    np.save(f"{unified_save_dir}/BQX_{test_mc0.size}_{test_mc0.Blocks}", X)
 
-plt.plot(test_mc0.T, M/test_mc0.spin, label="M")
-plt.title(f"Magnetization vs Temperature of {test_mc0.MAT_NAME} at size {test_mc0.size}x{test_mc0.size}")
-plt.savefig(f"{test_mc0.save_directory}/M_{test_mc0.size}.png")
-plt.close()
-plt.plot(test_mc0.T, X/X.max(), label="X")
-plt.title(f"Susceptibility vs Temperature of {test_mc0.MAT_NAME} at size {test_mc0.size}x{test_mc0.size}")
-plt.savefig(f"{test_mc0.save_directory}/X_{test_mc0.size}.png")
-plt.close()
-M_1, X_1 = M, X
 
[…]
