### 2025-03-12 #7589 @arkavo
Hi @tpurcell90 and @ChrisWoodgate

- The imports have been changed in the ```sample.py``` example.
- New (better) sample scripts have been added under the ```/sample_scripts``` (not under ```/src```). These should be portable enough that you can run it from anywhere. The inputs have also been added in the same folder (additional help in the next point)
- The MonteCarlo class now supports ```input_folder``` and ```output_folder``` args, in case you want to source/save at different places. The default is still the ```inputs/outputs``` folder in case you run ```sample.py```

The GPU driver and CUDA version

![Image](https://github.com/user-attachments/assets/9eab22b5-5735-4391-b9e0-ed087a06b0d2)

The environment can be replicated by ```pip``` installing
- pycuda
- montecarlo
- seaborn
- matplotlib
- tqdm

full image of installed packages

![Image](https://github.com/user-attachments/assets/b0147f08-f21a-4a54-83fd-eecdef1897d4)

Hopefully this all comes together

Thanks.
