### 2025-03-14 #7589 @tpurcell90
The sample.py script now runs for me, as an FYI there is a typo in the code package:

```
Traceback (most recent call last):
  File "/home/u9/purcellt/git/CUDA-METRO/testing/test2/sample.py", line 16, in <module>
    np.save(f"{test_mc0.save_directory}/grid_{i:04d}", test_mc0.run_mc_dmi_66612(test_mc0.T[0]))
```

Running more tests now
