### 2025-03-17 #7589 @arkavo
Hi @tpurcell90, are you sure you're on the latest version? because I specifically removed the "cannot find dmi6.py" issue (which was not needed anyways) from the code.

![Image](https://github.com/user-attachments/assets/6f84d4a9-913c-4d95-99e3-84abd543c931)

(right here)

The ```sample.py``` script *should* create a folder at ```/src/cudametro``` (working directory of the script) with spin values as ```.npy```. But of course, they cannot be accessed unless the "Analyze" part is up and running. Are the scripts in the ```sample_scripts``` folder running for you?

Thanks.
