### 2025-03-24 #7589 @tpurcell90
Hi @arkavo the sample script is now working as intended. I am not sure what happened.

The fig reproduction scripts are still giving memory errors. What needs to change for that?
