### 2025-03-24 #7589 @arkavo
Hi @tpurcell90, at this point I think it might be some incorrect bindings on your environment, but I'm not sure. I've used a Nvidia RTX 4070Ti to validate the scripts with no issues whatsoever. If you want to increase compatibility, please try decreasing the "Blocks" field in the config files (supplied in the folder itself). See if it runs with ```"Blocks" : "2" ``` first

Here are my library bindings for reference:

![Image](https://github.com/user-attachments/assets/9e14a2b0-05a7-4834-aa63-3ae29bb8d903)

(you can ignore the 11.8 bindings)

Hope this works.
Thanks.
