### 2025-03-24 #7589 @tpurcell90
Running with blocks=2 works, and my bindings are correct. It could be memory limitations on the system that is causing the problem as I am doing this on an HPC in interactive mode.
