### 2025-03-25 #7589 @tpurcell90
For the checklist the only remaining points is not having a full API, but a usage guide and no automated tests.

However, I am not sure what would be useful to test in this package.
