### 2025-03-27 #7589 @arkavo
@tpurcell90 by "full API" you mean definition for all functions in the code? or something else entirely. Also by automated tests, do you mean a check if the package is installed correctly? Sorry for the question, its my first time actually packaging a project.

I would say if you test out the 3 diagrams (sample_scripts), they cover the entire code-base one way or another
- 3 separate crystal types
- Analysis of the output files
Suggestions on this would be highly appreciated

Thanks.
