### 2025-03-27 #7589 @tpurcell90
Hi @arkavo yes so a full API would be essentially docstrings for all of the functions/classes in the code with everything defined. There is  a usage guide so that might not be necessary, but for some of the core functions it might also be good for there to be doc strings on them.

For automated tests what I mean is that I can run something like `pytest tests/` and be certain that 1) the code is installed properly and 2) any changes one might make does not break some other part of the code.

The exact nature of these tests can be different for different code bases, so what to test is up to the developer
