+++ src/cudametro/__init__.py
@@ -1,3 +1,5 @@
+__version__ = '1.2.1'
+
 import sys
 #import montecarlo as montecarlo
 from .construct import MonteCarlo
