+++ setup.py
@@ -13,8 +13,8 @@ except Exception:
 setup(
 	# Name of the package 
 	name='CUDA-METRO',
+	version="1.2.1",
 	packages=find_packages(),
-	version='1.2.0',
 	# Short description of your library 
 	description='pyCUDA Metropolis Monte Carlo 2D Heisenberg Model Simulation',
 	# Long description of your library 
