+++ src/cudametro/construct.py
@@ -19,7 +19,6 @@ from matplotlib.colors import Normalize
 import os 
 import sys 
 import re
-#import Material_Reader as rm
 script_dir = os.path.dirname(os.path.abspath(__file__))
 sys.path.insert(0, script_dir)
 import cudametro.montecarlo as mc
@@ -30,6 +29,9 @@ import json
 import datetime
 
 def read_2dmat(filename):
+    '''
+    Reads the 2D material file from the ../../Inputs folder
+    '''
     namelist = []
     params_list = np.array([])
     with open(filename, 'r') as f:
@@ -42,6 +44,9 @@ def read_2dmat(filename):
     return namelist, np.array(params_list, dtype=np.float32)
 
 class MonteCarlo:
+    '''
+    Default MC class, use this to instantiate the Simulation object
+    '''
     def __init__(self, config, input_folder="../../inputs/", output_folder="outputs/"):
         with open(config, 'r') as f:
             CONFIG = json.load(f)
@@ -95,7 +100,6 @@ class MonteCarlo:
         spin_gpu = np.array([self.spin]).astype(np.float32)
         self.SGPU = drv.mem_alloc(spin_gpu.nbytes)
         drv.memcpy_htod(self.SGPU, spin_gpu)
-        #self.save_directory = "../../"+self.Output_Folder+self.Prefix+"_"+self.Material+"_"+str(datetime.datetime.now().strftime("%Y_%m_%d_%H_%M_%S"))
         self.save_directory = "Output_"+self.Prefix+"_"+self.Material+"_"+str(datetime.datetime.now().strftime("%Y_%m_%d_%H_%M_%S"))
         self.NBS = int(self.MAT_PARAMS[20]), int(self.MAT_PARAMS[21]), int(self.MAT_PARAMS[22]), int(self.MAT_PARAMS[23])
         print("Output Folder default path: ", self.save_directory)
@@ -128,6 +132,9 @@ class MonteCarlo:
             f.write(metadata_file)
         
     def display_material(self):
+        '''
+        Print material properties
+        '''
         print(f"\tMaterial: {self.Material}")
         print(f"\tSize: {self.size}")
         print(f"\tSpin: {self.spin}")
@@ -141,7 +148,10 @@ class MonteCarlo:
         print(f"\tFile TC/NC: {self.MAT_PARAMS[24]}")
         print(f"\tConfig Temps: {self.Temps}")
         
-    def generate_random_numbers(self, mult):
+    def generate_random_numbers(self):
+        '''
+        Generate 4 uniform random number arrays for the simulation and copies them to the GPU
+        '''
         self.NLIST = rg.gen_uniform((self.C1), np.float32)
         self.ULIST = rg.gen_uniform((self.C1), np.float32)
         self.VLIST = rg.gen_uniform((self.C1), np.float32)
@@ -155,7 +165,10 @@ class MonteCarlo:
         mc.NPREC(self.NLIST, self.NFULL, self.GSIZE, block=(1,1,1), grid=(self.C1,1,1))
         mc.VPREC(self.ULIST, self.VLIST, self.S1FULL, self.S2FULL, self.S3FULL, self.SGPU, block=(1,1,1), grid=(self.C1,1,1))
 
-    def generate_ising_numbers(self, mult):
+    def generate_ising_numbers(self):
+        '''
+        Generate 4 uniform random number arrays for the simulation and copies them to the GPU
+        '''
         self.NLIST = rg.gen_uniform((self.C1), np.float32)
         self.ULIST = rg.gen_uniform((self.C1), np.float32)
[…]
