### 2025-03-28 4f9d168
ver_1.2.2 package upload, readme fixes, docstrings added
