+++ src/cudametro/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.2.1'
+__version__ = '1.2.2'
 
 import sys
 #import montecarlo as montecarlo
