### 2025-03-28 #7589 @arkavo
Hi @tpurcell90, I've added docstrings in the code and updated it to v1.2.2, please update before running the pytest scripts in the ```/tests``` folder.
@ChrisWoodgate do you have anything to add on this or any additional feedback?

Thanks
