### 2025-04-01 #7589 @phibeck
Thank you @tpurcell90 for completing your checklist! 🎉 If there are no further comments/suggestions for the author at this point, could you please confirm that you completed your checklist and agree to the publication of this software in its current form?

@ChrisWoodgate when you have a moment, could you provide an update on whether the issues you raised have been resolved? Thank you
