### 2025-04-02 #7589 @ChrisWoodgate
Hi @arkavo

Thanks for looking into this! 🙂

If I navigate to `/tests` and run `python3 test.py` I get the following output:

```
=== Testing the CUDAMETRO package ===
=== Testing the import of the necessary packages ===
 Searching for CUDAMETRO
Please install the package using the following command: pip install cudametro
 Searching for numpy
NumPy package is installed. Version: 2.2.4
 Searching for matplotlib
Matplotlib package is installed. Version: 3.10.1
 Searching for tqdm
TQDM package is installed. Version: 4.67.1
 Searching for seaborn
Seaborn package is installed. Version: 0.13.2
 Searching for pyCUDA
PyCUDA package is installed. Version: 2025.1
=== All packages are installed ===
=== Testing GPU availability ===
 Searching for GPU
GPU found: 1 devices found
Device 0: NVIDIA RTX A2000 12GB
=== GPU Information ===

Device 0: NVIDIA RTX A2000 12GB
Compute Capability: (8, 6)
Total Memory: 11881.9375 MB
Max Threads per Block: 1024
Max Threads per Multiprocessor: 1536
Max Grid Size: 2147483647, 65535, 65535
Max Threads per Block: 1024

CHECK COMPLETE
=== All tests are complete ===
=== Please refer to the README.md file for further instructions ===
=== If you encounter any issues, please refer to the documentation ===
=== For any further queries, please contact the author ===
=== Thank you for using the CUDAMETRO package ===
```

If I then run `python3 runtime_test.py` I get the output

```
=== Testing the CUDAMETRO package ===
Inputs Folder default path:
Output Folder default path:  Output_0_MnSTe_2025_04_02_14_28_46
<pycuda._driver.DeviceAllocation object at 0x148920a0cdc0>
	Material: MnSTe
	Size: 10
	Spin: 1.0
	J1: 10.5200 J2: 0.0000 J3: 0.0000 J4: 0.0000
	K1x: 0.0000 K1y: 0.0000 K1z: 0.0040
	K2x: 0.0000 K2y: 0.0000 K2z: 0.0000
	K3x: 0.0000 K3y: 0.0000 K3z: 0.0000
	K4x: 0.0000 K4y: 0.0000 K4z: 0.0000
	Ax: 0.0000 Ay: 0.0000 Az: 0.2900
	NBS System: (6, 6, 6, 12)
	File TC/NC: 140.0
	Config Temps: [np.float32(10.0)]
Stability Runs:   0%|                                                            | 0/50 [00:00<?, ?runs/s]
Error during Monte Carlo simulation: cuFuncSetBlockShape failed: an illegal memory access was encountered
0 files found.....Analyzing
Error initializing Analyze: cuMemAlloc failed: an illegal memory access was encountered
Error during visualization: 'NoneType' object has no attribute 'spin_view'
=== CUDAMETRO package test completed ===
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
```
