### 2025-04-02 #7589 @arkavo
Hi @ChrisWoodgate I saw the output and it says that

![Image](https://github.com/user-attachments/assets/c67dc495-7c9c-4c46-b39f-7c7b835091a9)
Maybe its an import issue overall?
Does your environment have the package? if not you can just simply drag it from pypi (```pip install cudametro```). The current version is 1.2.2
