### 2025-04-02 #7589 @tpurcell90
My apologies, but this conversation sparked me to take a closer look at this, and I tried it one more time and it randomly started to work with blocks=4 twice, but failed 8 or so more times.

Since it worked once this is not a device or environment bug as that did not change, and since it is an intermittent `an illegal memory access` my guess is something is getting overwritten in memory. Have you tried using a mem checker on the C code?

Some context that might help you debug. If one prints out the block line the failure rate decreases. And if it fails it always fails when `j=512` on the first iteration here
```python
'''
    DMI SECTOR
    This section has the MC simulation algorithms for materials that use a DMI term
    In order, this contains the following functions:
    run_mc_dmi_66612: Runs the MC simulation for a 6 6 6 12 crystal structure material
    run_mc_dmi_4448: Runs the MC simulation for a 4 4 4 8 crystal structure material
    run_mc_dmi_3636: Runs the MC simulation for a 3 6 3 6 crystal structure material
    '''
    def run_mc_dmi_66612(self, T):
        beta = np.array([1.0 / (T * 8.6173e-2)],dtype=np.float32)
        drv.memcpy_htod(self.BJ,beta[0])
        for j in range(self.stability_runs):
            mc.METROPOLIS_MC_DM1_6_6_6_12(
                self.GPU_MAT,
                self.GRID_GPU,
                self.BJ,
                self.NFULL[j*self.Blocks:(j+1)*self.Blocks-1],
                self.S1FULL[j*self.Blocks:(j+1)*self.Blocks-1],
                self.S2FULL[j*self.Blocks:(j+1)*self.Blocks-1],
                self.S3FULL[j*self.Blocks:(j+1)*self.Blocks-1],
                self.RLIST[j*self.Blocks:(j+1)*self.Blocks-1],
                self.GPU_TRANS,
                self.GPU_DMI_6,
                self.B_GPU,
                self.GSIZE,
                block=(self.Threads,1,1),
                grid=(self.Blocks,1,1)
            )
            mc.GRID_COPY(self.GRID_GPU, self.GPU_TRANS, block=(1,1,1), grid=(self.Blocks,1,1))
        drv.memcpy_dtoh(self.grid, self.GRID_GPU)
        return self.grid
```
