+++ src/cudametro/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.2.2'
+__version__ = '1.2.3'
 
 import sys
 #import montecarlo as montecarlo
