### 2025-04-03 #7589 @ChrisWoodgate
Hi @arkavo

Thanks for making those changes. I am now happy with the paper and have checked all the relevant boxes on my 'checklist'.

I am still having issues running the test scripts, though.

If I run `python3 test.py` I get this output, showing that everything is installed correctly:
```
=== Testing the CUDAMETRO package ===
=== Testing the import of the necessary packages ===
 Searching for CUDAMETRO
CUDA METRO package is installed. Version: 1.2.3
 Searching for numpy
NumPy package is installed. Version: 2.2.4
 Searching for matplotlib
Matplotlib package is installed. Version: 3.10.1
 Searching for tqdm
TQDM package is installed. Version: 4.67.1
 Searching for seaborn
Seaborn package is installed. Version: 0.13.2
 Searching for pyCUDA
PyCUDA package is installed. Version: 2025.1
=== All packages are installed ===
=== Testing GPU availability ===
 Searching for GPU
GPU found: 1 devices found
Device 0: NVIDIA RTX A2000 12GB
=== GPU Information ===

Device 0: NVIDIA RTX A2000 12GB
Compute Capability: (8, 6)
Total Memory: 11881.9375 MB
Max Threads per Block: 1024
Max Threads per Multiprocessor: 1536
Max Grid Size: 2147483647, 65535, 65535
Max Threads per Block: 1024

CHECK COMPLETE
=== All tests are complete ===
=== Please refer to the README.md file for further instructions ===
=== If you encounter any issues, please refer to the documentation ===
=== For any further queries, please contact the author ===
=== Thank you for using the CUDAMETRO package ===
```

Then if I run `python3 runtime_test.py` with `"Blocks"    : 2,` in `runtime_test_config.json` I get output as expected:
```
=== Testing the CUDAMETRO package ===
Inputs Folder default path:
Output Folder default path:  Output_0_MnSTe_2025_04_03_09_30_12
	Material: MnSTe
	Size: 10
	Spin: 1.0
	J1: 10.5200 J2: 0.0000 J3: 0.0000 J4: 0.0000
	K1x: 0.0000 K1y: 0.0000 K1z: 0.0040
	K2x: 0.0000 K2y: 0.0000 K2z: 0.0000
	K3x: 0.0000 K3y: 0.0000 K3z: 0.0000
	K4x: 0.0000 K4y: 0.0000 K4z: 0.0000
	Ax: 0.0000 Ay: 0.0000 Az: 0.2900
	NBS System: (6, 6, 6, 12)
	File TC/NC: 140.0
	Config Temps: [np.float32(10.0)]
Stability Runs: 100%|█████████████████████████████████████████████████| 50/50 [00:01<00:00, 41.07runs/s]
50 files found.....Analyzing
Processing grid_0049.npy mode: spin heatmap

Exiting spin view mode
Processing grid_0049.npy mode: quiver

Exiting quiver view mode
=== CUDAMETRO package test completed ===
```

But if I change to `"Blocks"    : 4,` in `runtime_test_config.json` I am encountering an illegal memory access again
```
=== Testing the CUDAMETRO package ===
Inputs Folder default path:
Output Folder default path:  Output_0_MnSTe_2025_04_03_09_31_27
	Material: MnSTe
	Size: 10
	Spin: 1.0
	J1: 10.5200 J2: 0.0000 J3: 0.0000 J4: 0.0000
	K1x: 0.0000 K1y: 0.0000 K1z: 0.0040
	K2x: 0.0000 K2y: 0.0000 K2z: 0.0000
	K3x: 0.0000 K3y: 0.0000 K3z: 0.0000
	K4x: 0.0000 K4y: 0.0000 K4z: 0.0000
	Ax: 0.0000 Ay: 0.0000 Az: 0.2900
	NBS System: (6, 6, 6, 12)
	File TC/NC: 140.0
	Config Temps: [np.float32(10.0)]
Stability Runs:   0%|                                                          | 0/50 [00:00<?, ?runs/s]
Error during Monte Carlo simulation: cuFuncSetBlockShape failed: an illegal memory access was encountered
0 files found.....Analyzing
Error initializing Analyze: cuMemAlloc failed: an illegal memory access was encountered
Error during visualization: 'NoneType' object has no attribute 'spin_view'
=== CUDAMETRO package test completed ===
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuMemFree failed: an illegal memory access was encountered
PyCUDA WARNING: a clean-up operation failed (dead context maybe?)
cuModuleUnload failed: an illegal memory access was encountered
```

For reference, here is the output of `nvidia-smi` on my machine:
```
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 570.124.06             Driver Version: 570.124.06     CUDA Version: 12.8     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  NVIDIA RTX A2000 12GB          Off |   00000000:01:00.0 Off |                  Off |
| 30%   36C    P8              6W /   70W |     596MiB /  12282MiB |      1%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+

+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|    0   N/A  N/A           16588      G   /usr/libexec/Xorg                       553MiB |
+-----------------------------------------------------------------------------------------+
```

Perhaps @tpurcell90 could see if he can reproduce this issue?
