### 2025-04-03 #7589 @arkavo
@ChrisWoodgate thanks for info. I ran this same code on ```Blocks 4``` and ran it without errors. Additionally I ran it with ```Blocks 400``` with no errors.....

![Image](https://github.com/user-attachments/assets/acd1354d-3010-491f-a1a5-cd36e39eb126)

I will also wait for @tpurcell90 to try and replicate the issue while I think of fixes I guess. Thanks for your comments on the paper.
