+++ src/cudametro/construct.py
@@ -162,8 +162,8 @@ class MonteCarlo:
         self.S2FULL = pycuda.gpuarray.zeros((self.C1), dtype=np.float32)
         self.S3FULL = pycuda.gpuarray.zeros((self.C1), dtype=np.float32)
 
-        mc.NPREC(self.NLIST, self.NFULL, self.GSIZE, block=(1,1,1), grid=(self.C1,1,1))
-        mc.VPREC(self.ULIST, self.VLIST, self.S1FULL, self.S2FULL, self.S3FULL, self.SGPU, block=(1,1,1), grid=(self.C1,1,1))
+        mc.NPREC(self.NLIST, self.NFULL, self.GSIZE, block=(2,1,1), grid=(self.C1,1,1))
+        mc.VPREC(self.ULIST, self.VLIST, self.S1FULL, self.S2FULL, self.S3FULL, self.SGPU, block=(2,1,1), grid=(self.C1,1,1))
 
     def generate_ising_numbers(self, deprecate):
         '''
@@ -179,13 +179,13 @@ class MonteCarlo:
         self.S2FULL = pycuda.gpuarray.zeros((self.C1), dtype=np.float32)
         self.S3FULL = pycuda.gpuarray.zeros((self.C1), dtype=np.float32)
 
-        mc.NPREC(self.NLIST, self.NFULL, self.GSIZE, block=(1,1,1), grid=(self.C1,1,1))
-        mc.ISING(self.ULIST, self.VLIST, self.S1FULL, self.S2FULL, self.S3FULL, self.SGPU, block=(1,1,1), grid=(self.C1,1,1))
+        mc.NPREC(self.NLIST, self.NFULL, self.GSIZE, block=(2,1,1), grid=(self.C1,1,1))
+        mc.ISING(self.ULIST, self.VLIST, self.S1FULL, self.S2FULL, self.S3FULL, self.SGPU, block=(2,1,1), grid=(self.C1,1,1))
     
     def sampler(self):
         self.N_SAMPLE = rg.gen_uniform((self.Blocks), np.float32)
         self.GPU_N_SAMPLE = drv.mem_alloc(self.N_SAMPLE.nbytes)
-        mc.NPREC(self.N_SAMPLE, self.GPU_N_SAMPLE, self.GSIZE, block=(1,1,1), grid=(self.Blocks,1,1))
+        mc.NPREC(self.N_SAMPLE, self.GPU_N_SAMPLE, self.GSIZE, block=(2,1,1), grid=(self.Blocks,1,1))
 
     def mc_init(self):
         '''
@@ -238,24 +238,28 @@ class MonteCarlo:
             mc.METROPOLIS_MC_DM1_6_6_6_12(self.GPU_MAT, self.GRID_GPU, self.BJ, self.NFULL[j*self.Blocks:(j+1)*self.Blocks], self.S1FULL[j*self.Blocks:(j+1)*self.Blocks], self.S2FULL[j*self.Blocks:(j+1)*self.Blocks], self.S3FULL[j*self.Blocks:(j+1)*self.Blocks], self.RLIST[j*self.Blocks:(j+1)*self.Blocks], self.GPU_TRANS, self.GPU_DMI_6, self.B_GPU, self.GSIZE, block=(self.Threads,1,1), grid=(self.Blocks,1,1))
             mc.GRID_COPY(self.GRID_GPU, self.GPU_TRANS, block=(2,1,1), grid=(self.Blocks,1,1))
         drv.memcpy_dtoh(self.grid, self.GRID_GPU)
+        drv.Context.synchronize()
         return self.grid
+
 
     def run_mc_dmi_4448(self, T):
[…]
