### 2025-04-03 #7589 @tpurcell90
@arkavo It might be good to try a clean install of this on a different computer to see if you can reproduce the errors. Maybe you have something in the environment/old installation libraries that is avoiding this memory bug.
