### 2025-04-03 #7589 @arkavo
@tpurcell90 ,@ChrisWoodgate sorry for the ping but I'm sort of stuck. I've tried to replicate your conditions with *fully fresh install* of all libraries with 0 bindings which I used later. I'm even using a workstation card right now (new computer essentially, with only the NVIDIA-HPC stack pre-installed)
```
=== Testing the CUDAMETRO package ===
=== Testing the import of the necessary packages ===
 Searching for CUDAMETRO
CUDA METRO package is installed. Version: 1.2.3
 Searching for numpy
NumPy package is installed. Version: 2.2.4
 Searching for matplotlib
Matplotlib package is installed. Version: 3.10.1
 Searching for tqdm
TQDM package is installed. Version: 4.67.1
 Searching for seaborn
Seaborn package is installed. Version: 0.13.2
 Searching for pyCUDA
PyCUDA package is installed. Version: 2025.1
=== All packages are installed ===
=== Testing GPU availability ===
 Searching for GPU
GPU found: 1 devices found
Device 0: Tesla V100S-PCIE-32GB
=== GPU Information ===
```

with ```Blocks: 2``` working as intended. I however went up to ```Blocks: 800``` and *still* face no error

```
(cmtest) arkavo@nsdrl101v:~/CUDA-METRO/tests$ compute-sanitizer python runtime_test.py
========= COMPUTE-SANITIZER
=== Testing the CUDAMETRO package ===
Inputs Folder default path:
Output Folder default path:  Output_0_MnSTe_2025_04_04_01_38_02
        Material: MnSTe
        Size: 10
        Spin: 1.0
        J1: 10.5200 J2: 0.0000 J3: 0.0000 J4: 0.0000
        K1x: 0.0000 K1y: 0.0000 K1z: 0.0040
        K2x: 0.0000 K2y: 0.0000 K2z: 0.0000
        K3x: 0.0000 K3y: 0.0000 K3z: 0.0000
        K4x: 0.0000 K4y: 0.0000 K4z: 0.0000
        Ax: 0.0000 Ay: 0.0000 Az: 0.2900
        NBS System: (6, 6, 6, 12)
        File TC/NC: 140.0
        Config Temps: [np.float32(10.0)]
Stability Runs: 100%|█████████████████████████████████████| 50/50 [01:18<00:00,  1.58s/runs]
50 files found.....Analyzing
Processing grid_0049.npy mode: spin heatmap

Exiting spin view mode
Processing grid_0049.npy mode: quiver

Exiting quiver view mode
=== CUDAMETRO package test completed ===
========= ERROR SUMMARY: 0 errors
```

I was thinking that it was something that converted all fp64 into fp32 on the workstation card but that too does not appear to be the case. I will continue looking into it while the rest of the paper is going on.

Thank you for bearing with me.
