### 2025-04-03 #7589 @tpurcell90
Version 1.2.4 is not on the github repo with a git fetch.

If I run the test multiple times in a row I get this one in every ten or so attempts. Which makes debugging more difficult.
```
compute-sanitizer --tool memcheck python runtime_test.py
========= COMPUTE-SANITIZER
=== Testing the CUDAMETRO package ===
Inputs Folder default path:
Output Folder default path:  Output_0_MnSTe_2025_04_03_14_59_37
	Material: MnSTe
	Size: 10
	Spin: 1.0
	J1: 10.5200 J2: 0.0000 J3: 0.0000 J4: 0.0000
	K1x: 0.0000 K1y: 0.0000 K1z: 0.0040
	K2x: 0.0000 K2y: 0.0000 K2z: 0.0000
	K3x: 0.0000 K3y: 0.0000 K3z: 0.0000
	K4x: 0.0000 K4y: 0.0000 K4z: 0.0000
	Ax: 0.0000 Ay: 0.0000 Az: 0.2900
	NBS System: (6, 6, 6, 12)
	File TC/NC: 140.0
	Config Temps: [10.0]
Stability Runs: 100%|████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 50/50 [00:23<00:00,  2.13runs/s]
50 files found.....Analyzing
Processing grid_0049.npy mode: spin heatmap

Exiting spin view mode
Processing grid_0049.npy mode: quiver

Exiting quiver view mode
=== CUDAMETRO package test completed ===
========= ERROR SUMMARY: 0 errors
```
