### 2025-04-04 #7589 @tpurcell90
@arkavo and @ChrisWoodgate what compiler are your nvcc's wrapped against?

Mine is GCC 13.2.0, maybe something got update in the std lib that's causing this?
