### 2025-04-04 #7589 @ChrisWoodgate
@tpurcell90 @arkavo

I am using GCC `13.3.0` on my machine. Could support the suggestion from @tpurcell90 that maybe something has changed?
