### 2025-04-04 #7589 @ChrisWoodgate
@tpurcell90 @arkavo

One other thought did occur to me last night. Are we absolutely sure that the algorithm ('Algorithm 2' in the paper) avoids a data race? Is it possible for two CUDA threads to try to flip the same spin at the same time? From the way it is written in the paper, I think it _does_, because the we only update all spins at the end of the `for` loop.

It could be worth just double-checking this as a data race would lead to unpredictable behaviour.
