### 2025-04-04 #7589 @arkavo
@ChrisWoodgate the algo goes as follows:
- you pick a spin (from the initial state $S_0$)
- change it
- record the change (don't put it back yet)
- put in all the recorded changes at once (if a point is in 2 instances, it just picks one at random) $\rightarrow S_1$

In essence, all these changes see $S_0$, which it treats as an entirely static state. Data race conditions will not appear here, which works well for the present results. It would have mattered if we tried to put back the changes while making changes in some site which would probably return an access error(?).
