### 2025-04-04 #7589 @tpurcell90
A data race condition could describe  what @ChrisWoodgate and I are seeing, and if it is appearing because of some compiler optimization introduced after GCC 12.3 that would describe why @arkavo can't reproduce it. I discounted that previously because @arkavo never found it.

The errors in that thread all seem to be compile time so we should have seen them before runtime I would think.

One test would be to force the compilation to use "-O0" for all of us to see what happens. I'm not sure how to do this on CUDA but there must be a way somehow, but I don't know it off the top of my head.
