### 2025-04-04 #7589 @arkavo
@tpurcell90 (and @ChrisWoodgate) , I've (probably) succeeded in compiling the cpp code in ```options=["-std=c++17", "-O0"]```. These are bundled in v1.2.5, I've also forced cpp standards to ```c++17```.

Maybe this solves the issue ?

Thanks
