### 2025-04-05 #7589 @tpurcell90
On the older system and using GCC 8.3 I don't see any errors (tried ten times).

At this point I would be willing to accept that this is an error with pycuda using gcc-13 compilers causing this issue. The error appears in the block initialization step and seems to be compiler dependent.. And maybe document that somewhere in your readme there seems to be this instability?

Realistically the only way I think you could confirm this @arkavo it is to try using gcc-13 yourself.

@phibeck is there a policy for how to deal with this type of situation?
