### 2025-04-05 #7589 @arkavo
Additionally I found this on SO

![Image](https://github.com/user-attachments/assets/bee2c9df-627d-4126-9f2a-c2a661413cec) which reinforces your point.

I installed gcc-13.0 on my system and it does not *compile* at all throwing me

![Image](https://github.com/user-attachments/assets/3a0feb99-02cd-4cfa-8316-55b56c41317d)

but when I removed the symlinks

![Image](https://github.com/user-attachments/assets/76f60443-d1fb-4518-a132-8e5a78bf753b)

it starts working again perfectly.

I will wait for @ChrisWoodgate's opinion on the matter before committing to a solution. Is that fine with you @tpurcell90 ?
