### 2025-04-05 #7589 @ChrisWoodgate
Okay, sorry both for the slow reply. I am trying compiling/running the package on an older HPC system (with an older version of GCC) to which I have access but have to go via the queueing system to access GPU capabilities there (on the other machine I can run interactively) so it might take a while for me to run the tests.

I agree that with @arkavo that the algorithm in and of itself shouldn't cause a data race.

I also agree with @tpurcell90 's suggestion that one way to proceed would be for @arkavo to try compiling with gcc-13 to see if he can reproduce the error.

If I can't make the package work on this older HPC system then I am running out of options. All the GPU capacity to which I have access is on managed HPC systems so I don't have too much control over precisely which version of compilers I am using.

@phibeck , is there a JOSS-approved way to proceed here if I can't make the package work on my hardware? Currently I am unable to verify functionality of the package or reproducibility of the results in the manuscript.
