### 2025-04-05 #7589 @tpurcell90
I was able to reproduce everything using the older system. The only suggestion I'd make is set the figure 1 script to use tighter convergence criteria to better line up with the paper.

@arkavo I would be willing to test the updated one on my system to see if it works.
