### 2025-04-06 #7589 @arkavo
@tpurcell90 nice to hear that. I originally used double the size and quadruple the cores needed to run figure 1. Since we were having "out of memory" issues I made it easier to run but it of course gives us some difference with the results in the paper. I'll put it in the original numbers again and you can check (no need to install the package, a simple ```git pull``` will do)

As for gcc-13 @ChrisWoodgate, on my system it refuses to compile with gcc-13.1.0 as I mentioned before but works again as soon as I remove it, it might be because my pycuda (and entire NVIDIA stack) was *built* on another version of gcc? I'll let you know if I find anything else.

I've included the pycuda sourcemodule to use no optimization and ```std c++17``` while compiling the cpp part of the code. Hopefully that will alleviate some of the issues. An instruction on *not* using the ```gcc-13``` or above should be decent for the ```README``` since I'm also looking at forums and they say that ```gcc-13``` does not compile CUDA code properly:

https://forums.developer.nvidia.com/t/nvcc-12-3-with-gcc-13-2-not-working/271476

Looking forward to replies,
Thanks.
