### 2025-04-06 d588281
fig1 script update
