### 2025-04-06 #7589 @tpurcell90
> As for gcc-13 @ChrisWoodgate, on my system it refuses to compile with gcc-13.1.0 as I mentioned before but works again as soon as I remove it, it might be because my pycuda (and entire NVIDIA stack) was built on another version of gcc? I'll let you know if I find anything else.

Normally you need to recompile the Nvidia stack if you update compilers because of library mismatches.

For the figures I had to manually increase the point density for T to get it work (red curve missed the peak. I will try editing the stability_wrap number as well.
