### 2025-04-07 #7589 @phibeck
Thank you all for the updates, and special thanks to @ChrisWoodgate and @tpurcell90  for putting in all the work to help @arkavo figure out what's causing this somewhat erratic issue. I haven't come across such a case before, so I'll check in with the other editors and get back to you later today on how to proceed.
