### 2025-04-07 #7589 @tpurcell90
The suggested fix for the noise worked, so I am now happy with everything, outside of the compiler problems.
