### 2025-04-08 #7589 @phibeck
Hi All, after discussing with the editors I've decided that it would be good to dig a little deeper to narrow down the issue. Unless it's a known reported bug with the updated compiler, it remains more likely to be some undefined behavior in the program. At the same time @ChrisWoodgate isn't able to verify functionality, so I will add another reviewer. If we still end up with the same conclusion, we can proceed as suggested to document the issue in the README.

I briefly discussed the issue with my former colleague @the-hampel whose expertise lies within this scope. He has agreed to help out as a reviewer.
