+++ src/cudametro/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.2.5'
+__version__ = '1.2.6'
 
 import sys
 #import montecarlo as montecarlo
