### 2025-04-09 #7589 @the-hampel
Hi @arkavo,

I found the paper to be written very nicely and the code is also very accessible.

I have a few comments that I separated into code and paper review.

### code review

- [ ] In the installation requirements pycuda pinned to an old version 2024.1, why? With python 3.13 one needs the newer pycuda 2025.1? Is there a specific reason why only pycuda 2024.1 is supported? I found that pycuda 2025.1 seems to be also okay? Especially since pycuda 2024.1 does not support the latest cuda 12.8 version
- [ ] The [[readme.md](http://readme.md/)](http://readme.md/) is misleading. Please change the clone command to ssh instead of the deprecated https: git clone [[git@github.com](mailto:git@github.com)](mailto:git@github.com):arkavo/CUDA-METRO.git
- [ ] One cannot run the tests in `test/` without installing the code first. Please add an instruction to run `pip install ./` in the main directory of the repo to install it first before running `python runtime_test.py` Or change the test scripts so that they can run directly from the test directory before installing the code.
- [ ] I am not sure why the [`[sample.py](http://sample.py/)`](http://sample.py) is in the src/cudametro directory and not with the other examples in `sample_scripts` ?
- [ ] There is a problem in this cuda kernel function: https://github.com/arkavo/CUDA-METRO/blob/d58828114531894f9ca2853d4f72b78d88b3df06/src/cudametro/montecarlo.py#L991 : This is a non void function, that indicates it returns an int but there seems to be a branch in which the function does not return anything. Please provide a default return value to avoid exceptions.
- [ ] Upon running the test I see the same error for (with cuda 12.8 and gcc 13.3 installed through conda):

```fortran
Error during Monte Carlo simulation: cuFuncSetBlockShape failed: an illegal memory access was encountered
```

There is no indication that there is a problem with the installation I made, I think. Interestingly it works when I run the code with sanitizers:

```fortran
compute-sanitizer --tool memcheck python runtime_test.py
```

. But I think this should give you  the possibility to reproduce the problem? I tried also with an older cuda 12.4 version that ships in conda with gcc 12.4 (`conda install cuda==12.4.1` ) and I have the same problem.

The fact that the code suddenly works when running via the cuda memcheck feature could point to a problem within the code or maybe pycuda. Not necessarily to a wrong installation or incompatibility with gcc 13.3 . But I am also not 100% certain.

With the `--tool initcheck`  I actually find usage of unitialized variables:

```fortran
         Host Frame: function_call in driver.py:502
         Host Frame: run_mc_dmi_66612 in construct.py:239
         Host Frame: <module> in runtime_test.py:28

 Uninitialized __global__ memory read of size 4 bytes
     at cp_grid+0x60
     by thread (0,0,0) in block (19,0,0)
```

maybe it would be worth checking these? In particular in the kernel function for copying:

```fortran
__global__ void cp_grid(float_t* grid, float_t* tf)
{
    int idx = blockIdx.x;
    int threadID = idx;
    grid[int(tf[threadID*4])*3] = tf[threadID*4+1];
    grid[int(tf[threadID*4])*3+1] = tf[threadID*4+2];
    grid[int(tf[threadID*4])*3+2] = tf[threadID*4+3];
}
```

is there maybe a problem that this writes outside of the grid variable bounds? I did not analyze the code carefully but maybe this is causing trouble? Maybe the authors could clarify this. Anyway, maybe this is a dead end but worth checking in my opinion.

I think it should be fairly easy to reproduce the problem we are seeing by using coda to install cuda:

```fortran
conda create —name pycuda
conda activate pycuda
conda install python==3.13
conda install cuda -c nvidia
pip install pycuda numpy tqdm seaborn
```

Could you try exactly those steps on one of your systems and let us know if you can reproduce the error? Then we can take it from there.

### paper review:

- [ ] The following sentence is a bit misleading:
”It was once thought that long-range magnetic order could not exist (Mermin & Wagner, 1966)
in two or one-dimensional materials…”
One could read this sentence in a way that you say that the Mermin-Wagner theorem was proven to be wrong. But what Huang et al. demonstrated is that in 2D system with anisotropy there is long-range order, which is totally to be expected theoretically since the full rotational symmetry is reduced and M&W theorem is only about continuous symmetry breaking not discrete ones. I would appreciate if you could write the sentence in a way that it is clear that there have been 2D systems found that due to anisotropy no longer fall into the Mermin & Wagner theorem.

Also could you clarify for me if your proposed MC algorithm is entirely new, or based on an idea already published? This was not entirely clear to me. I understood that it is completely new, and thus I wonder if you plan to publish the algorithm itself somewhere else? Maybe it could be highlighted a bit better in the paper if this is a completely new idea on how to implement the MC on a GPU for this hamiltonian. Thanks!

I hope these comments will improve the code and the paper!

Best,
Alex
