### 2025-04-10 #7589 @arkavo
Hey @the-hampel, thanks for the fast feedback.

- I've built another system from scratch with your *exact conda environment* with ```GCC-14.2.1``` and ```CUDA-12.8``` on an RTX 3050. I can now reproduce your error (crashing on default, runs with ```memcheck```, shows warning on ```initcheck```)
- I've also used pyCuda 2025.1 for this installation (with CUDA 12.8)
- the ```sample.py``` is designed to be run from inside its directory, it draws the functions from the local files ```construct.py```, hence it isn't out there.
- I'm looking into the copy function but till now I haven't found any obvious errors.

I will fix the paper/README issues as we go

Thanks.
