+++ src/cudametro/montecarlo.py
@@ -63,10 +63,12 @@ __global__ void cp_grid(float_t* grid, float_t* tf)
 {
     int idx = blockIdx.x;
     int threadID = idx;
-    grid[int(tf[threadID*4])*3] = tf[threadID*4+1];
-    grid[int(tf[threadID*4])*3+1] = tf[threadID*4+2];
-    grid[int(tf[threadID*4])*3+2] = tf[threadID*4+3];
+    int site = __float2uint_rz(tf[threadID*4]);
+    grid[site*3] = tf[threadID*4+1];
+    grid[site*3+1] = tf[threadID*4+2];
+    grid[site*3+2] = tf[threadID*4+3];
 }
+/*
 __global__ void alt_cp_grid(float_t* grid, float_t* tf)
 {
     int idx = blockIdx.x;
@@ -75,6 +77,7 @@ __global__ void alt_cp_grid(float_t* grid, float_t* tf)
     grid[int(tf[threadID*4])*4+1] = tf[threadID*4+2];
     grid[int(tf[threadID*4])*4+2] = tf[threadID*4+3];
 }
+*/
 //Vector preprocessing
 __global__ void uvec_processor(float_t* u, float_t* v, float_t* s1, float_t* s2, float_t* s3, float_t* spin)
 {
@@ -674,14 +677,14 @@ __global__ void metropolis_mc_dm1_6_6_6_12(float_t *mat, float_t *sheet, float_t
     
     if (dE < 0)
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
     }
     else if (expf(-1.0*dE*T[0]) > R[threadID])
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
@@ -711,14 +714,14 @@ __global__ void metropolis_mc_dm0_6_6_6_12(float_t *mat, float_t *sheet, float_t
     
     if (dE < 0)
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
     }
     else if (expf(-1.0*dE*T[0]) > R[threadID])
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
@@ -748,14 +751,14 @@ __global__ void metropolis_mc_dm0_3_6_3_6(float_t *mat, float_t *sheet, float_t 
 
     if (dE < 0)
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
     }
     else if (expf(-1.0*dE*T[0]) > R[threadID])
     {
-        tf[threadID*4] = pt_thread;
+        tf[threadID*4] = (float_t)pt_thread;
         tf[threadID*4+1] = S1[threadID];
         tf[threadID*4+2] = S2[threadID];
         tf[threadID*4+3] = S3[threadID];
[…]
