### 2025-04-10 #7589 @the-hampel
Hi @arkavo ,

that looks very promising. I just tested the new version and I can confirm that for me the errors are gone using cuda 12.8 and gcc 13.3. So it seemed that the grid copy indeed did something wrong? I am not 100% sure I understand the commit but I am happy that this pointed you to the right function.

Once you made revisions in the paper and readme I will have a look.

Best,
Alex
