+++ references.bib
@@ -358,3 +358,49 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
   year      =  2020,
   copyright = "http://creativecommons.org/licenses/by/4.0/"
 }
+
+@ARTICLE{Gong2017-br,
+  title     = "Discovery of intrinsic ferromagnetism in two-dimensional van der
+               Waals crystals",
+  author    = "Gong, Cheng and Li, Lin and Li, Zhenglu and Ji, Huiwen and
+               Stern, Alex and Xia, Yang and Cao, Ting and Bao, Wei and Wang,
+               Chenzhe and Wang, Yuan and Qiu, Z Q and Cava, R J and Louie,
+               Steven G and Xia, Jing and Zhang, Xiang",
+  abstract  = "The realization of long-range ferromagnetic order in
+               two-dimensional van der Waals crystals, combined with their rich
+               electronic and optical properties, could lead to new magnetic,
+               magnetoelectric and magneto-optic applications. In
+               two-dimensional systems, the long-range magnetic order is
+               strongly suppressed by thermal fluctuations, according to the
+               Mermin-Wagner theorem; however, these thermal fluctuations can
+               be counteracted by magnetic anisotropy. Previous efforts, based
+               on defect and composition engineering, or the proximity effect,
+               introduced magnetic responses only locally or extrinsically.
+               Here we report intrinsic long-range ferromagnetic order in
+               pristine Cr2Ge2Te6 atomic layers, as revealed by scanning
+               magneto-optic Kerr microscopy. In this magnetically soft,
+               two-dimensional van der Waals ferromagnet, we achieve
+               unprecedented control of the transition temperature (between
+               ferromagnetic and paramagnetic states) using very small fields
+               (smaller than 0.3 tesla). This result is in contrast to the
+               insensitivity of the transition temperature to magnetic fields
+               in the three-dimensional regime. We found that the small applied
+               field leads to an effective anisotropy that is much greater than
+               the near-zero magnetocrystalline anisotropy, opening up a large
+               spin-wave excitation gap. We explain the observed phenomenon
+               using renormalized spin-wave theory and conclude that the
+               unusual field dependence of the transition temperature is a
+               hallmark of soft, two-dimensional ferromagnetic van der Waals
+               crystals. Cr2Ge2Te6 is a nearly ideal two-dimensional Heisenberg
+               ferromagnet and so will be useful for studying fundamental spin
+               behaviours, opening the door to exploring new applications such
+               as ultra-compact spintronics.",
+  journal   = "Nature",
+  publisher = "Springer Science and Business Media LLC",
+  volume    =  546,
+  number    =  7657,
+  pages     = "265--269",
+  month     =  jun,
+  year      =  2017,
+  language  = "en"
+}
