### 2025-04-10 #7589 @arkavo
Hi @the-hampel,

To address some of your points:

Code review

- Development of this project began in 2024, so we were using the python 2024 version. Since now we've checked that 2025.1 is working, I will update the ```requirements.pip``` with *minimum* requirements instead of exact.
- Changed from git clone https to ssh
- I've added a text in the README that advises the user to install the package before running the tests
- Solved
- These were deprecated functions, I've commented them out.
- Solved (hopefully)

Paper review:

- Changed the structure of the sentence to better reflect the advancements in the field and how the introduction of anisotropy brings forth the "not falling into Mermin Wagner" category of materials. We *specifically* use this anisotropy in this code.
- This algorithm is new to the best of our knowledge and I've not seen mentions of something like this elsewhere. As for publishing a standalone of this, I've never really thought about it but I'm open to options.

The Bug:

I'll just address the bug in brief here: it turns out that just *allocating* memory on the GPU is not enough, it has to be populated once by some matrix to properly make it work, else it just throws "illegal address" errors. The target for the copy function, a small sub array was memory allocated but not "initialized" per se, once I solved that by making this sub array some shape of ```np.zeros(np.float32)```, the GCC compiler started behaving. I've tested this across 4 different OS, GPU and Compiler versions and they all seem to work.

@tpurcell90 and @ChrisWoodgate is there anything you guys want to ask/clarify/want me to add ?

Thanks.

PS: No need to update the package, this recent commit is just some README and paper stuff
