+++ paper.md
@@ -105,7 +105,7 @@ Unlike the SSU scheme, we do not extract simulation data after every $N^2$ spin 
 
 We validate our algorithm by accurately reproducing temperature-dependent magnetic phase transitions and intricate topological spin textures already reported in the literature, either by SSU based Monte Carlo or by Landau Lifshitz Gilbert(LLG) [@gilbert_classics_2004] spin dynamics, for diverse crystal structures.
 
-It was once thought that long-range magnetic order could not exist [@PhysRevLett.17.1133] in two or one-dimensional materials. In 2017, $\text{CrI}_\text{3}$ was one of the first 2D materials [@Huang2017] where ferromagnetism was demonstrated. In their seminal work from 1967, Mermin and Wagner[@PhysRevLett.17.1133] used the isotropic Heisenberg model to analytically demonstrate that long-range order could not exist in materials with less than three dimensions. Nonetheless, in 2017 two independent research groups experimentally proved the existence of long-range ferromagnetism at low finite temperatures in two distinct 2D crystals: $\text{CrI}_\text{3}$ [@Huang2017] and $\text{Cr}_\text{2}\text{Ge}_\text{2}\text{Te}_\text{6}$ [@Gong2017-br]. A new frontier in nanotechnology has eventually emerged, since multiple other research groups have shown that 2D magnetism can exist in various materials, even at room temperature. Finite-size effects and magneto-crystalline anisotropy, which Mermin and Wagner did not consider, are thought to be connected to the formation of 2D magnetism. To accurately simulate the magnetic states of two-dimensional materials,the Heisenberg model must be augmented with necessary anisotropic terms. Using CUDA-METRO, we first simulate the magnetic phase transition of $\text{CrI}_\text{3}$ from ferromagnetic to paramagnetic with increasing temperature. In Fig 1, we reproduce magnetic phase transition from [@Kartsev2020], and show the effect of parallelization with respect to the reference data. As mentioned before, we can obtain very accurate results with respect to SSU methods by limiting the parallelization at $10\%$ or below of $N^2$. The ferromagnetic to paramagnetic transition point is known as Critical temperature (or the Curie point) and is extracted from the peak of the susceptibility versus temperature plot.
+In their seminal work from 1967, Mermin and Wagner[@PhysRevLett.17.1133] used the isotropic Heisenberg model to analytically demonstrate that long-range order could not exist in materials with less than three dimensions. Nonetheless, in 2017 two independent research groups experimentally proved the existence of long-range ferromagnetism at low finite temperatures in two distinct 2D crystals: $\text{CrI}_\text{3}$ [@Huang2017] and $\text{Cr}_\text{2}\text{Ge}_\text{2}\text{Te}_\text{6}$ [@Gong2017-br]. A new frontier in nanotechnology has eventually emerged, since multiple other research groups have shown that 2D magnetism can exist in various materials, even at room temperature. Finite-size effects and magneto-crystalline anisotropy, which Mermin and Wagner did not consider, are thought to be connected to the formation of 2D magnetism. To accurately simulate the magnetic states of two-dimensional materials,the Heisenberg model must be augmented with necessary anisotropic terms. Using CUDA-METRO, we first simulate the magnetic phase transition of $\text{CrI}_\text{3}$ from ferromagnetic to paramagnetic with increasing temperature. In Fig 1, we reproduce magnetic phase transition from [@Kartsev2020], and show the effect of parallelization with respect to the reference data. As mentioned before, we can obtain very accurate results with respect to SSU methods by limiting the parallelization at $10\%$ or below of $N^2$. The ferromagnetic to paramagnetic transition point is known as Critical temperature (or the Curie point) and is extracted from the peak of the susceptibility versus temperature plot.
 
[…]
