### 2025-04-10 #7589 @the-hampel
Hi @arkavo ,

thank you for addressing all my questions so quickly. I am very happy with the current form of the paper and all my questions related to the repository are addressed. I hope @tpurcell90 and @ChrisWoodgate can confirm that the code now works for them as well in the environment they originally found the problem / error.

Very happy that this improved the stability of the code!

Best regards,
Alex
