### 2025-04-17 #7589 @ChrisWoodgate
No worries, @arkavo

I am just getting started on this. The good news is that everything appears to be running as intended for me now—thanks to @the-hampel and @tpurcell90 for their work diagnosing the issue, and to @arkavo for making the necessary changes.

I now just need to progress through the scripts for figure reproduction one by one and check that I can reproduce the results of the paper. I will be heading out of the office soon (this weekend is a long holiday weekend in the UK because of Easter) but I hope to have this completed either over the weekend or early next week.
