### 2025-04-23 #7589 @ChrisWoodgate
Okay, thanks for bearing with me while I got around to working on this, @phibeck . I have worked through the three scripts for reproducing the paper's figures this afternoon and a confirm that I reproduce the results satisfactorily for Figures 1 and 2.

For Fig. 3, I stack into an out of memory error, though, which I assume is because the script is trying to allocate more memory than the GPU has available!

My output from `fig3_script.py` is

```
150 files found.....Analyzing
Processing grid_0149.npy mode: spin heatmap

Exiting spin view mode
Processing grid_0149.npy mode: quiver

Exiting quiver view mode
Traceback (most recent call last):
  File "/home/epp/phuqtp/joss_reviews/CUDA-METRO/sample_scripts/fig3_script.py", line 39, in <module>
    viewer.en_66612()
  File "/home/epp/phuqtp/.local/lib/python3.9/site-packages/cudametro/construct.py", line 655, in en_66612
    mc.EN_CALC_6_6_6_12(self.GPU_MAT, self.GRID_GPU, self.B_GPU, self.GSIZE, GPU_ET, self.GPU_DMI_6, block=(2,1,1), grid=(self.Blocks,1,1))
  File "/home/epp/phuqtp/.local/lib/python3.9/site-packages/pycuda/driver.py", line 502, in function_call
    func._launch_kernel(grid, block, arg_buf, shared, None)
pycuda._driver.LaunchError: cuLaunchKernel failed: too many resources requested for launch
```

while the output of `nvidia-smi` on my machine is

```
Wed Apr 23 18:03:22 2025
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 570.133.20             Driver Version: 570.133.20     CUDA Version: 12.8     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  NVIDIA RTX A2000 12GB          Off |   00000000:01:00.0 Off |                  Off |
| 30%   42C    P8             11W /   70W |     596MiB /  12282MiB |      0%      Default |
|                                         |                        |                  N/A |
+-----------------------------------------+------------------------+----------------------+

+-----------------------------------------------------------------------------------------+
| Processes:                                                                              |
|  GPU   GI   CI              PID   Type   Process name                        GPU Memory |
|        ID   ID                                                               Usage      |
|=========================================================================================|
|    0   N/A  N/A            6970      G   /usr/libexec/Xorg                       553MiB |
+-----------------------------------------------------------------------------------------+
```

Remind me—how much memory did you have available on your GPU, @arkavo ? Were you able to reproduce Fig. 3, @tpurcell90 @the-hampel ? If so, I may have to concede that my lowly RTX A2000 is not enough to reproduce the final figure!
