### 2025-04-23 #7589 @ChrisWoodgate
Great, thanks @arkavo!

I am now running into some weird glitch with my Python install. (My machine rebooted this afternoon and it seems to have 'lost' some of the Matplotlib backend during the reboot—don't ask me how!) This is entirely on me to solve, but it is late here now so I will look again tomorrow. 🙂
