### 2025-04-24 #7589 @the-hampel
Hi @arkavo ,

I tried now as well ( I have to admit I did not try fig3 before). I ran into the same problem as @ChrisWoodgate and I am more and more confused.

I tried to understand the config.json file and found that the documentation is a bit sparse on that end. I ended up with this config file:
```
{
    "Single_Mat_Flag" : 1,
    "Animation_Flags" : 0,
    "DMI_Flag"        : 1,
    "TC_Flag"         : 0,
    "Static_T_Flag"   : 1,
    "FM_Flag"         : 0,
    "Input_flag"      : 0,
    "Input_File"      : "Starting_States/state.npy",
    "Temps"           : ["10.0"],
    "Material"        : "MnSTe",
    "Multiple_Materials" : "TCS",
    "SIZE"    : 3,
    "Box"       : 4,
    "Blocks"    : 128,
    "Threads"   : 2,
    "B"         : "1.5",
    "stability_runs"    : 100,
    "calculation_runs"  : 40,
    "stability_wrap"    : 50,
    "calculation_wrap"  : 35,
    "Cmpl_Flag"         : 1,
    "Prefix"            : "0"
}
```
and it keeps crashing with:
```
Traceback (most recent call last):
  File "/scratch/hampel/CUDA-METRO/sample_scripts/fig3_script.py", line 39, in <module>
    viewer.en_66612()
    ~~~~~~~~~~~~~~~^^
  File "/fsc/home/hampel/micromamba/envs/pycuda/lib/python3.13/site-packages/cudametro/construct.py", line 659, in en_66612
    mc.EN_CALC_6_6_6_12(self.GPU_MAT, self.GRID_GPU, self.B_GPU, self.GSIZE, GPU_ET, self.GPU_DMI_6, block=(2,1,1), grid=(self.Blocks,1,1))
    ~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
  File "/fsc/home/hampel/micromamba/envs/pycuda/lib/python3.13/site-packages/pycuda/driver.py", line 502, in function_call
    func._launch_kernel(grid, block, arg_buf, shared, None)
    ~~~~~~~~~~~~~~~~~~~^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
pycuda._driver.LaunchError: cuLaunchKernel failed: too many resources requested for launch
```
This is now a really small grid I think and I am running this on 2x A30 with 24GB of VRAM, and during running the calculation it only uses a few 100 Mbyte:
```
╰─❯ nvidia-smi
Thu Apr 24 09:35:35 2025
+-----------------------------------------------------------------------------------------+
| NVIDIA-SMI 550.54.14              Driver Version: 550.54.14      CUDA Version: 12.4     |
|-----------------------------------------+------------------------+----------------------+
| GPU  Name                 Persistence-M | Bus-Id          Disp.A | Volatile Uncorr. ECC |
| Fan  Temp   Perf          Pwr:Usage/Cap |           Memory-Usage | GPU-Util  Compute M. |
|                                         |                        |               MIG M. |
|=========================================+========================+======================|
|   0  NVIDIA A30                     On  |   00000000:81:00.0 Off |                    0 |
| N/A   32C    P0             28W /  165W |       0MiB /  24576MiB |      0%      Default |
|                                         |                        |             Disabled |
+-----------------------------------------+------------------------+----------------------+
|   1  NVIDIA A30                     On  |   00000000:A1:00.0 Off |                    0 |
| N/A   32C    P0             29W /  165W |       0MiB /  24576MiB |      0%      Default |
|                                         |                        |             Disabled |
+-----------------------------------------+------------------------+----------------------+
```
 I am more suspecting that it asks for too many threads per block or something? But I am not sure how to control this?

I also tried to just run the post-processing as you suggested but with the same problem.

On what system are you running these successfully?

Best,
Alex
