+++ src/cudametro/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '1.2.7'
+__version__ = '1.2.8'
 
 import sys
 #import montecarlo as montecarlo
