### 2025-04-24 #7589 @the-hampel
Yes now it seems to work for me without problems. The resulting plots look good to me. Thanks @arkavo
