### 2025-04-24 #7589 @ChrisWoodgate
Yup, it works for me now, too. Thanks for sorting this, @arkavo

I have re-run the test suite and the three sample scripts and am now satisfied that I can reproduce the results of the paper. I'll complete my reviewer checklist now
