### 2025-04-24 #7589 @arkavo
Hi @phibeck,

- I can confirm that the authors/affiliations/ORCIDs are correct.
- v1.2.8 is the latest release which the reviewers @tpurcell90 @the-hampel and @ChrisWoodgate have graciously helped me fix and release.
- I have at least uploaded the archive on Zenodo with the link ```https://zenodo.org/records/15275575```

![Image](https://github.com/user-attachments/assets/937a0b00-5a26-44dc-a4ae-164f7f36f010)

- The title and author list with ORCID and affiliations match on Zenodo

![Image](https://github.com/user-attachments/assets/f1960d98-71aa-4d46-bfdb-f2a699f761f7)

- License is fine (Apache 2.0)

If I am missing any requirement or did something incorrectly, please tell me.

Thanks.
