+++ references.bib
@@ -396,6 +396,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
                behaviours, opening the door to exploring new applications such
                as ultra-compact spintronics.",
   journal   = "Nature",
+  DOI       = {https://doi.org/10.1038/nature22060},
   publisher = "Springer Science and Business Media LLC",
   volume    =  546,
   number    =  7657,
