### 2025-04-24 a8fda6c
Update references.bib
Updated 2 DOIs
