+++ references.bib
@@ -314,6 +314,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
                   will briefly review these developments as a tribute to Robin
                   Bullough who was a great admirer of the LLG equation.",
   journal      = "arXiv [nlin.PS]",
+  DOI          = {10.48550/arXiv.1101.1005},
   publisher    = "arXiv",
   year         =  2011,
   primaryClass = "nlin.PS"
@@ -396,7 +397,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
                behaviours, opening the door to exploring new applications such
                as ultra-compact spintronics.",
   journal   = "Nature",
-  DOI       = {https://doi.org/10.1038/nature22060},
+  DOI       = {10.1038/nature22060},
   publisher = "Springer Science and Business Media LLC",
   volume    =  546,
   number    =  7657,
