### 2025-05-05 #7589 @phibeck
Thanks. Two more things:

- could you clarify the role of the second author, since there do not seem to be any commits by them?
- could you make sure the title in the archive matches that of the JOSS paper? Thanks
