### 2025-05-05 #7589 @arkavo
Hi @phibeck,
                The second author is my supervisor who framed the problem statement and secured funding. I wrote the entire code by myself and generated/analyzed data. Both authors had equal contribution in writing the paper.

                Where would this "author contributions" lie? At the end of the paper in a new section?

Thanks.
