### 2025-05-05 e78d6be
Updated Paper and Refs
