+++ references.bib
@@ -83,7 +83,7 @@
 }
 
 @article{augustin_properties_2021,
-	title = {Properties and dynamics of meron topological spin textures in the two-dimensional magnet {CrCl3}},
+	title = {Properties and dynamics of meron topological spin textures in the two-dimensional magnet $\text{CrCl}_\text{3}$},
 	volume = {12},
 	copyright = {2021 The Author(s)},
 	issn = {2041-1723},
@@ -216,7 +216,7 @@
 
 @article{acs_nanolett,
 author = {Cui, Qirui and Zhu, Yingmei and Ga, Yonglong and Liang, Jinghua and Li, Peng and Yu, Dongxing and Cui, Ping and Yang, Hongxin},
-title = {Anisotropic Dzyaloshinskii–Moriya Interaction and Topological Magnetism in Two-Dimensional Magnets Protected by P4m2 Crystal Symmetry},
+title = {Anisotropic Dzyaloshinskii-Moriya Interaction and Topological Magnetism in Two-Dimensional Magnets Protected by $P\bar{4}m2$ Crystal Symmetry},
 journal = {Nano Letters},
 volume = {22},
 number = {6},
@@ -246,7 +246,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
 }
 
 @article{du_spontaneous_2022,
-	title = {Spontaneous Magnetic Skyrmions in Single-Layer $CrInX_3$ ({X} = {Te}, {Se})},
+	title = {Spontaneous Magnetic Skyrmions in Single-Layer $\text{CrInX}_text{3}$ ({X} = {Te}, {Se})},
 	volume = {22},
 	copyright = {https://doi.org/10.15223/policy-029},
 	issn = {1530-6984, 1530-6992},
