+++ references.bib
@@ -216,7 +216,7 @@
 
 @article{acs_nanolett,
 author = {Cui, Qirui and Zhu, Yingmei and Ga, Yonglong and Liang, Jinghua and Li, Peng and Yu, Dongxing and Cui, Ping and Yang, Hongxin},
-title = {Anisotropic Dzyaloshinskii-Moriya Interaction and Topological Magnetism in Two-Dimensional Magnets Protected by $P\bar{4}m2$ Crystal Symmetry},
+title = {Anisotropic {Dzyaloshinskii}-{Moriya} Interaction and Topological Magnetism in Two-Dimensional Magnets Protected by $P\bar{4}m2$ Crystal Symmetry},
 journal = {Nano Letters},
 volume = {22},
 number = {6},
@@ -263,7 +263,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
 }
 
 @article{PhysRevLett.17.1133,
-  title = {Absence of Ferromagnetism or Antiferromagnetism in One- or Two-Dimensional Isotropic Heisenberg Models},
+  title = {Absence of Ferromagnetism or Antiferromagnetism in One- or Two-Dimensional Isotropic {Heisenberg} Models},
   author = {Mermin, N. D. and Wagner, H.},
   journal = {Phys. Rev. Lett.},
   volume = {17},
@@ -278,7 +278,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
 }
 
 @article{Huang2017,
-  title = {Layer-dependent ferromagnetism in a van der Waals crystal down to the monolayer limit},
+  title = {Layer-dependent ferromagnetism in a van der {Waals} crystal down to the monolayer limit},
   volume = {546},
   ISSN = {1476-4687},
   url = {http://dx.doi.org/10.1038/nature22391},
@@ -362,7 +362,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
 
 @ARTICLE{Gong2017-br,
   title     = "Discovery of intrinsic ferromagnetism in two-dimensional van der
-               Waals crystals",
+               {Waals} crystals",
   author    = "Gong, Cheng and Li, Lin and Li, Zhenglu and Ji, Huiwen and
                Stern, Alex and Xia, Yang and Cao, Ting and Bao, Wei and Wang,
                Chenzhe and Wang, Yuan and Qiu, Z Q and Cava, R J and Louie,
