### 2025-05-05 #7589 @arkavo
Hi @phibeck,
                  I fixed the paper according to the suggestions. The equation however, probably cannot be shortened efficiently, since it dot products $s_i$ and $s_j$ in the $J_1$ term but does $s_i$ and $s_k$ in $J_2$ term and so on. This will probably confuse the reader even more if I try to condense it. I will think of something better in the meantime however. The other things should be fixed and I've also changed the language/phrasing in the paper wherever suggested. I am unsure about whether I can fix capitalization in references but I've tried anyways. I will compile the paper here and give you a thumbs up when its ready.

Thanks.
