### 2025-05-13 #7589 @phibeck
Hi @arkavo, sorry for taking so long, things are just very busy at the moment. Thanks for making the changes. There are a few more comments:

- line 27: "finding _the_ ground state _of_"
- line 32: "ground state _of_ ultra large"..
- line 43f: regarding the equation, okay. If you prefer to keep it as is, that's okay. But in my opinion the current version is more confusing, because the name of the dummy indices doesn't mean anything generally. But I leave this up to you.
- line 62: there is another occurrence of k_b, and please use text font for B (since it denotes Boltzmann)
- line 71f: Thanks for rephrasing, now I understand it better. To further simplify the sentence, could you adjust it to: "These 4 random number arrays are further divided into an array depicting the site location 𝑛, the angle coordinates for a new random spin vector 𝜃 and 𝜙, and a conditional uniform random number 𝑟, which is used to evaluate the Metropolis criteria."
- line 82: add whitespace between sentences
- line 116: "which is an emergent _phenomenon_ in ..."
- line 124f: I have to admit I am still not sure I understand the meaning of the sentence "An antiferromagnetic skyrmion spin texture is accurately reproduced in our simulation which are technologically important since they do not exhibit skyrmion Hall effect." Do you mean something like: "An antiferromagnetic skyrmion spin texture is accurately reproduced in our simulation, which is technologically significant due to the absence of a skyrmion Hall effect in such systems."
- line 171: can you check the formatting again for CrInX_3? There's an issue

Thank you also for adjusting the archive title. Would you mind removing the hyphen in Monte Carlo to match the paper title after the corrections? Thank you.
