+++ references.bib
@@ -246,7 +246,7 @@ eprint = {https://doi.org/10.1021/acs.nanolett.1c04803}
 }
 
 @article{du_spontaneous_2022,
-	title = {Spontaneous Magnetic Skyrmions in Single-Layer $\text{CrInX}_text{3}$ ({X} = {Te}, {Se})},
+	title = {Spontaneous Magnetic Skyrmions in Single-Layer $\text{CrInX}_\text{3}$ ({X} = {Te}, {Se})},
 	volume = {22},
 	copyright = {https://doi.org/10.15223/policy-029},
 	issn = {1530-6984, 1530-6992},
