+++ paper.md
@@ -58,7 +58,7 @@ $$
 $$
 Where $J$ is the isotropic exchange parameter, the $K$s are the anisotropic exchange parameters, with the superscript denoting the spin direction, $A$ is the single ion exchange parameter, $\lambda$ is the biquadratic parameter, $D$ is the Dyzaloshinskii-Moriya Interaction (DMI) parameter. $\mu$ is the dipole moment of a single atom and $B$ is the external magnetic field. $s_i,s_j$ are individual atomic spin vectors. $\{s_j\}$ are the first set of neighbours, $\{s_k\}$ are the second set of neighbours and so on. The subscripts below all $J$s and $K$s denote the neighbour set, $J_1$ denotes the first neighbours, $J_2$ the second and so on. In our code, we have limited the number of neighbour sets to be 4 since it is expected for 2D materials that the interaction energy dies down beyond that. All these above parameters except $B$ are material specific parameters that are the inputs to our MC code.
 
-Starting from a random spin configuration, our objective is to find the orientation of spin vectors for every atom for the ground state of the lattice system for a given magnetic field and temperature. Traditionally, single spin update (SSU) scheme is employed to solve this problem, which satisfies a detailed balance condition. In the SSU method of updating the state, a single atomic spin is chosen at random and changed, while noting down the energy shift. This new state is then accepted or rejected using the Metropolis criteria as shown in Algorithm 1, where $\beta=(k_\text{B}T)^{-1}$, $k_b$ being the Boltzmann constant and $T$ being the temperature. It is clear that SSU becomes extremely inefficient as the dimensionality increases.
+Starting from a random spin configuration, our objective is to find the orientation of spin vectors for every atom for the ground state of the lattice system for a given magnetic field and temperature. Traditionally, single spin update (SSU) scheme is employed to solve this problem, which satisfies a detailed balance condition. In the SSU method of updating the state, a single atomic spin is chosen at random and changed, while noting down the energy shift. This new state is then accepted or rejected using the Metropolis criteria as shown in Algorithm 1, where $\beta=(k_\text{B}T)^{-1}$, $k_\text{B}$ being the Boltzmann constant and $T$ being the temperature. It is clear that SSU becomes extremely inefficient as the dimensionality increases.
 \begin{algorithm}
     \caption{Metropolis Selection}
     \label{algorithm:MS1}
