### 2025-05-13 #7589 @arkavo
Hi @phibeck,

- line 27: done
- line 32: done
- line 43f: Yeah I would prefer to keep it as it is. I'm sure the reader will notice the pattern. I tried shortening it once but the variable indices on the spins made it more confusing.
- line 62: done
- line 71f: done (thanks for the idea)
- line 82: done
- line 116: done
- line 124f: I _do_ mean it like that, thanks for the idea again.
- line 171: done

The hyphen in the archive should also be fixed. Thanks for looking over the paper text.

Thanks.
