### 2025-05-14 #7589 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
