### 2025-05-17 #7589 @Kevin-Mattheus-Moerman
- [ ] The paper uses both British English (e.g. `neighbours`, `neighbour`, `centred-rectangular`) and American English (e.g. `normalized`, `stabilize`, `parallelization`), please edit so one is used consistently.
- [ ] Check if `neucleation` should read `nucleation`
