### 2025-05-17 3e495a4
Update paper.md
Converted all text to British English and fixed 2 spell errors
