### 2025-05-17 #7589 @arkavo
Hi @Kevin-Mattheus-Moerman,

The issues should be fixed. I've also compiled the paper, when you've got time please give it a look

Thanks
