### 2025-05-17 47aaf23
Update paper.md
minor spelling mistake
