+++ paper.md
@@ -104,7 +104,7 @@ At present, five different lattice types (square, rectangular, centred-rectangul
 For a lattice of size $N\times N$, $100\%$ parallelisation would correspond to selecting $N^2$ spins at random.
 Since each spin selection and its consequent Metropolis criterion is evaluated on a separate CUDA core, it becomes apparent that we would need $N^2$ CUDA cores to achieve this $100\%$ parallelisation. Since the proposed algorithm may not adhere to the SSU detailed balance conditions, it yields approximate results, and there is a trade-off between parallelisation/acceleration and accuracy. It is found that if the parallelisation is limited to $10\%$ of the lattice size, we obtain very accurate results with significant acceleration.
 
-Unlike the SSU scheme, we do not extract simulation data after every $N^2$ spin change, rather we let it run till all the CUDA cores have sychronized and arrived at the next block check, which defines our step. To directly compare, if the SSU has $1$ spin changed between each step, our algorithm will have $P$ amount of spins changed per iteration, where $P$ is also the number of CUDA cores used or parallelisation used. In the conventional SSU, data is extracted at every $N^2$ (a full sweep) steps, compared to which, we would take $\frac{N^2}{P}$ steps to reach the same data point.
+Unlike the SSU scheme, we do not extract simulation data after every $N^2$ spin change, rather we let it run till all the CUDA cores have synchronise and arrived at the next block check, which defines our step. To directly compare, if the SSU has $1$ spin changed between each step, our algorithm will have $P$ amount of spins changed per iteration, where $P$ is also the number of CUDA cores used or parallelisation used. In the conventional SSU, data is extracted at every $N^2$ (a full sweep) steps, compared to which, we would take $\frac{N^2}{P}$ steps to reach the same data point.
 
 We validate our algorithm by accurately reproducing temperature-dependent magnetic phase transitions and intricate topological spin textures already reported in the literature, either by SSU based Monte Carlo or by Landau Lifshitz Gilbert (LLG) [@gilbert_classics_2004] spin dynamics, for diverse crystal structures.
 
