### 2025-05-19 72b10e9
Create CITATION.cff
