+++ README.md
@@ -2,6 +2,9 @@
 
 [![PyPI version](https://badge.fury.io/py/cudametro.svg)](https://badge.fury.io/py/cudametro)
 
+HTML: <a href="https://joss.theoj.org/papers/a7010c1182f894d9a781ff2d44036783"><img src="https://joss.theoj.org/papers/a7010c1182f894d9a781ff2d44036783/status.svg"></a>
+Markdown: [![status](https://joss.theoj.org/papers/a7010c1182f894d9a781ff2d44036783/status.svg)](https://joss.theoj.org/papers/a7010c1182f894d9a781ff2d44036783)
+
 # Description/Installation
 
 A pyCUDA based Metropolis Monte Carlo simulator for 2D-magnetism systems. This code uses NVIDIA CUDA architecture wrapped by a simple python wrapper to work.
